refactor(types,data,functions,scripts,firestore): Utils/Data/Types領域を機能別モジュールに分割
- types/index.ts (422行) → 10個の機能別モジュール + バレルexport
- data/dev-stories/episodes.ts (1709行) → エピソード別6ファイル + バレル
- functions/src/index.ts (561行) → helpers, ocr-schedule, tasting-analysis
- scripts/update-version.ts (502行) → types, version-logic, utils, file-updaters
- lib/firestore/userData.ts (349行) → write-queue, crud + バレル

既存のimportパスは全て維持（バレルエクスポートで吸収）

Closes #93
