+++ data/dev-stories/episode-002.ts
@@ -1,0 +1,162 @@
+import type { DevStoryEpisode } from '@/types';
+
+export const EPISODE_002: DevStoryEpisode = {
+  id: 'episode-002',
+  title: 'バイブコーディングとは？',
+  subtitle: 'AIと一緒にアプリを作る新しい方法',
+  imageUrl: '/dev-stories/vibe_coding.png',
+  dialogues: [
+    {
+      id: 'msg-002-1',
+      characterId: 'dori',
+      content: '「バイブコーディング」っていう言葉を聞いたんだけど、どんなものなのかな？',
+    },
+    {
+      id: 'msg-002-2',
+      characterId: 'server',
+      content: 'それはね、AIと会話しながらアプリを作っていく新しい方法だよ。',
+    },
+    {
+      id: 'msg-002-3',
+      characterId: 'dori',
+      content: '会話でアプリが作れるなんて、なんだか不思議だな...どういう仕組みなのかな？',
+    },
+    {
+      id: 'msg-002-4',
+      characterId: 'server',
+      content: '「こんな機能が欲しいな」って言葉で伝えると、AIがプログラムのコードを書いてくれるんだ。',
+    },
+    {
+      id: 'msg-002-5',
+      characterId: 'dori',
+      content: 'ということは...プログラミング言語を覚えなくても大丈夫なのかな？',
+    },
+    {
+      id: 'msg-002-6',
+      characterId: 'server',
+      content: 'そうだね。だから「バイブ」＝雰囲気やイメージで伝えるっていう意味なんだよ。',
+    },
+    {
+      id: 'msg-002-7',
+      characterId: 'dori',
+      content: 'そうなんだ...このローストプラスもそうやって作られたのかな？',
+    },
+    {
+      id: 'msg-002-8',
+      characterId: 'server',
+      content: '実はね、ローストプラスは、ほぼ全部バイブコーディングで作られたんだよ。',
+    },
+    {
+      id: 'msg-002-9',
+      characterId: 'dori',
+      content: 'このアプリ全部が...！それはすごいことだね。',
+    },
+    {
+      id: 'msg-002-10',
+      characterId: 'server',
+      content: '「担当表をシャッフルする機能が欲しい」みたいに伝えるだけで、形になっていくんだ。',
+    },
+    {
+      id: 'msg-002-11',
+      characterId: 'dori',
+      content: 'どんなAIツールを使っているのかな？',
+    },
+    {
+      id: 'msg-002-12',
+      characterId: 'server',
+      content: 'メインはCursorっていうツールで、他にもCodex、Antigravity、最近はClaude Codeにも挑戦中なんだ。',
+    },
+    {
+      id: 'msg-002-13',
+      characterId: 'dori',
+      content: 'そんなにたくさんあるんだ...使い分けているのかな？',
+    },
+    {
+      id: 'msg-002-14',
+      characterId: 'server',
+      content: 'うん。ツールによって得意なことが違うから、場面に合わせて選んでいるよ。',
+    },
+    {
+      id: 'msg-002-15',
+      characterId: 'dori',
+      content: 'とても便利そうだけど...難しいこともあるのかな？',
+    },
+    {
+      id: 'msg-002-16',
+      characterId: 'server',
+      content: 'あるね。AIに意図を正確に伝えるのが難しいし、思った通りにならないことも多いんだ。\n何度もやり直す根気が必要かもね。',
+    },
+    {
+      id: 'msg-002-17',
+      characterId: 'server',
+      content: 'あと、AIが書いたコードを理解せずに使うと、セキュリティの問題が起きることもあるよ。\nちゃんと確認することが大切だね。',
+    },
+    {
+      id: 'msg-002-18',
+      characterId: 'dori',
+      content: '便利だけど、ちゃんと向き合う姿勢が大事なんだね。',
+    },
+    {
+      id: 'msg-002-19',
+      characterId: 'server',
+      content: 'そうだね。でも挑戦する価値は十分あると思うよ。\nアイデアを形にする敷居がぐっと下がったからね。',
+    },
+  ],
+  detailContent: `バイブコーディングは、AIとの対話でアプリを作る新しいプログラミング手法です。
+
+## バイブコーディングとは
+
+**バイブコーディング**とは、プログラミング言語を使わず、自然な言葉でAIに「こんな機能が欲しい」と伝えてコードを生成してもらう開発方法です。
+
+「バイブ（Vibe）」は英語で「雰囲気」や「感覚」という意味。細かい技術的な指示ではなく、**やりたいことのイメージを伝えるだけ**でアプリが作れることから、この名前がつきました。
+
+## 従来のプログラミングとの違い
+
+| 従来のプログラミング | バイブコーディング |
+|---------------------|-------------------|
+| プログラミング言語を習得する必要がある | 自然な言葉で伝えるだけ |
+| コードを1行ずつ書く | AIがコードを生成 |
+| エラーを自分で解決 | AIと一緒に解決 |
+| 学習に長い時間がかかる | すぐに始められる |
+
+## メリット
+
+- **プログラミング経験がなくてもアプリが作れる**
+- **アイデアを素早く形にできる**
+- **試行錯誤のスピードが格段に上がる**
+
+## ローストプラスで使っているAIツール
+
+| ツール | 特徴 | このアプリでの使い方 |
+|--------|------|---------------------|
+| **Cursor** | AIエディタのパイオニア | メインの開発ツールとして日常的に使用 |
+| **Codex** | OpenAI製・高い推論力 | 複雑なバックエンド機能の修正 |
+| **Antigravity** | 最新のAI系IDE | フロントエンド改善、シャッフルロジック改善 |
+| **Claude Code** | ターミナルベースの対話型 | 最近挑戦中の新しいツール |
+
+それぞれのツールには得意分野があり、場面に合わせて使い分けています。
+
+## デメリット・注意点
+
+バイブコーディングは便利ですが、難しさや注意点もあります：
+
+- **適切な指示を出す難しさ**: AIに意図を正確に伝えるスキルが必要
+- **根気が必要**: 思い通りにならないことも多く、何度もやり直すことがある
+- **評価力が必要**: AIの提案が正しいかどうか判断する力が求められる
+- **セキュリティリスク**: 理解せずにコードを採用すると、危険な場合がある
+
+**大切なのは、AIに任せきりにせず、一緒に作っていく姿勢です。**
+
+## このアプリでの実例
+
+ローストプラスでは、以下の機能がバイブコーディングで作られました：
+
+- 担当表シャッフル機能
+- 試飲感想記録機能
+- この「開発秘話」ページ自体も！
+
+非エンジニアでも、アイデアがあれば形にできる時代になりました。`,
+  tags: ['バイブコーディング', 'AI', '開発手法'],
+  publishedAt: '2026-01-16',
+  order: 2,
+};
