+++ data/dev-stories/episode-003.ts
@@ -1,0 +1,289 @@
+import type { DevStoryEpisode } from '@/types';
+
+export const EPISODE_003: DevStoryEpisode = {
+  id: 'episode-003',
+  title: 'ローストプラスは一度作り直している',
+  subtitle: 'FlutterからNext.jsへの技術移行',
+  imageUrl: '/dev-stories/flutter_to_nextjs.png',
+  dialogues: [
+    {
+      id: 'msg-003-1',
+      characterId: 'kettle',
+      content: 'ローストプラスって、最初はFlutterで作ってたんだろ？',
+    },
+    {
+      id: 'msg-003-2',
+      characterId: 'mill',
+      content: 'そうなんだ。Flutterっていうのは、1つのコードでiPhoneとAndroid両方のアプリが作れるフレームワーク（開発ツール）なんだよね。',
+    },
+    {
+      id: 'msg-003-3',
+      characterId: 'kettle',
+      content: '便利じゃん！でも今はFlutterじゃないんだろ？何があった？',
+    },
+    {
+      id: 'msg-003-4',
+      characterId: 'mill',
+      content: '実は本番環境で大きな問題が起きてね...\nログインがうまくいかなくなったっていうわけ。',
+    },
+    {
+      id: 'msg-003-5',
+      characterId: 'kettle',
+      content: 'ログインできなくなった？それヤバいじゃん！',
+    },
+    {
+      id: 'msg-003-6',
+      characterId: 'mill',
+      content: 'そうなんだ。開発中のテスト環境では動いてたのに、実際にユーザーが使う本番環境では動かなくなっちゃって...',
+    },
+    {
+      id: 'msg-003-7',
+      characterId: 'kettle',
+      content: '原因はわかったのか？',
+    },
+    {
+      id: 'msg-003-8',
+      characterId: 'mill',
+      content: '正直、完全には特定できなかったんだ。\n何度も修正を試みたけど、直してもまた別の問題が出てきてね...',
+    },
+    {
+      id: 'msg-003-9',
+      characterId: 'kettle',
+      content: 'もぐらたたき状態か。それは厄介だな。',
+    },
+    {
+      id: 'msg-003-10',
+      characterId: 'mill',
+      content: 'まさにそう！\nFlutterとFirebase（データを保存するサービス）の認証の組み合わせが複雑で、どこで問題が起きてるか追跡しづらかったんだ。',
+    },
+    {
+      id: 'msg-003-11',
+      characterId: 'kettle',
+      content: 'で、作り直す決断をしたわけだ？',
+    },
+    {
+      id: 'msg-003-12',
+      characterId: 'mill',
+      content: 'そうなんだ。修正し続けるより、最初から作り直した方が早いし確実だと判断したっていうわけ。\n苦渋の決断だったけどね。',
+    },
+    {
+      id: 'msg-003-13',
+      characterId: 'kettle',
+      content: '新しく選んだNext.jsってのは何だ？',
+    },
+    {
+      id: 'msg-003-14',
+      characterId: 'mill',
+      content: 'Next.jsは、ウェブアプリを作るためのフレームワークなんだ。\nReact（リアクト）っていう有名な技術をベースにしてるんだよ。',
+    },
+    {
+      id: 'msg-003-15',
+      characterId: 'kettle',
+      content: 'なんでFlutterじゃなくてNext.jsにしたんだ？',
+    },
+    {
+      id: 'msg-003-16',
+      characterId: 'mill',
+      content: 'いくつか理由があるんだ。\nまず、ウェブ技術の方が情報が多くて、問題が起きても解決しやすいっていうわけ。',
+    },
+    {
+      id: 'msg-003-17',
+      characterId: 'mill',
+      content: 'それと、Flutterはアプリストアからダウンロードが必要だけど、ウェブアプリならURLにアクセスするだけで使えるんだ。',
+    },
+    {
+      id: 'msg-003-18',
+      characterId: 'kettle',
+      content: 'アプリダウンロード不要ってのはいいな！手軽だ。',
+    },
+    {
+      id: 'msg-003-19',
+      characterId: 'mill',
+      content: 'それにね、Next.jsとFirebaseの組み合わせは認証の仕組みがシンプルで、トラブルが起きにくいんだよ。',
+    },
+    {
+      id: 'msg-003-20',
+      characterId: 'kettle',
+      content: '他にも選択肢はあったのか？',
+    },
+    {
+      id: 'msg-003-21',
+      characterId: 'mill',
+      content: 'もちろん！React Native（リアクトネイティブ）とか、Flutter Webとか色々検討したんだ。\nでも、ローストプラスの使い方を考えると、ウェブアプリが一番合ってたっていうわけ。',
+    },
+    {
+      id: 'msg-003-22',
+      characterId: 'kettle',
+      content: 'ユーザーにとって何が変わった？',
+    },
+    {
+      id: 'msg-003-23',
+      characterId: 'mill',
+      content: 'まず、ログインが安定したんだ！\n以前みたいに「あれ、ログインできない...」っていうストレスがなくなったっていうわけ。',
+    },
+    {
+      id: 'msg-003-24',
+      characterId: 'kettle',
+      content: 'それは重要だな！他には？',
+    },
+    {
+      id: 'msg-003-25',
+      characterId: 'mill',
+      content: 'ページの読み込みも速くなったんだ。\nNext.jsは表示速度の最適化が得意なフレームワークなんだよ。',
+    },
+    {
+      id: 'msg-003-26',
+      characterId: 'kettle',
+      content: 'でも作り直すのは大変だっただろ？',
+    },
+    {
+      id: 'msg-003-27',
+      characterId: 'mill',
+      content: '正直、かなり大変だったんだ。\n一からデザインを考え直して、機能も全部実装し直したからね。',
+    },
+    {
+      id: 'msg-003-28',
+      characterId: 'mill',
+      content: 'でも、バイブコーディングのおかげで思ったより早く形になったんだ。\nAIと一緒に開発できる時代でよかったよ。',
+    },
+    {
+      id: 'msg-003-29',
+      characterId: 'kettle',
+      content: 'この失敗から学んだことは何だ？',
+    },
+    {
+      id: 'msg-003-30',
+      characterId: 'mill',
+      content: 'たくさんあるんだ。\nまず、本番環境でのテストをもっと早い段階でやるべきだったっていうこと。',
+    },
+    {
+      id: 'msg-003-31',
+      characterId: 'mill',
+      content: 'それと、技術選定は「作りやすさ」だけじゃなく「問題が起きたときの対処しやすさ」も大事だって学んだんだ。',
+    },
+    {
+      id: 'msg-003-32',
+      characterId: 'kettle',
+      content: '失敗を次に活かしてるじゃん。それが大事だ。',
+    },
+    {
+      id: 'msg-003-33',
+      characterId: 'mill',
+      content: 'ユーザーの皆さんには迷惑をかけてしまって申し訳なかったけど、結果的により良いアプリになったと思うんだ。',
+    },
+    {
+      id: 'msg-003-34',
+      characterId: 'kettle',
+      content: '素直に失敗認めて改善する。それができてるなら問題ないだろ。',
+    },
+    {
+      id: 'msg-003-35',
+      characterId: 'mill',
+      content: 'ありがとう。これからも安定して使えるアプリを目指して頑張るよ！\n使ってくれるみんなに感謝なんだ。',
+    },
+  ],
+  detailContent: `ローストプラスは、一度Flutterで開発した後、Next.jsで作り直しています。この記事では、その背景と経緯を詳しく説明します。
+
+## Flutterで何が起きたのか
+
+### 本番環境でのログイン問題
+
+Flutterで開発したアプリをリリースした後、**本番環境でログインができなくなる問題**が発生しました。
+
+- 開発中のテスト環境では正常に動作していた
+- 実際のユーザーが使う本番環境でのみ問題が発生
+- ログイン処理が途中で止まったり、認証状態が保持されなかったりした
+
+### 原因の特定困難性
+
+問題の根本原因を完全に特定することはできませんでした。考えられる要因として：
+
+- FlutterとFirebase Authenticationの複雑な連携
+- 本番環境と開発環境の微妙な設定差異
+- 認証状態の管理に関する複数のレイヤー
+
+**何度も修正を試みましたが、1つ直すと別の問題が発生するという「もぐらたたき」状態**に陥りました。
+
+## なぜNext.jsを選んだのか
+
+### フレームワーク選定の理由
+
+| 比較項目 | Flutter | Next.js |
+|----------|---------|---------|
+| **情報量** | 比較的少ない | 非常に多い |
+| **認証連携** | 複雑 | シンプル |
+| **インストール** | アプリストア経由 | URL直接アクセス |
+| **更新反映** | ストア審査が必要 | 即時反映可能 |
+| **デバッグ** | やや困難 | ブラウザで容易 |
+
+### 他の選択肢との比較
+
+以下のフレームワークも検討しました：
+
+- **React Native**: アプリ形式を維持できるが、同様の認証問題リスク
+- **Flutter Web**: ウェブ対応だが、モバイル向け設計が影響
+- **Vue.js / Nuxt**: 良い選択肢だが、エコシステムの規模でNext.jsを選択
+
+**ローストプラスの主な使用シーン（店舗での利用）を考慮し、インストール不要なウェブアプリが最適と判断**しました。
+
+### 作り直しの決断
+
+修正を続けるか、作り直すかの判断基準：
+
+- ✅ 修正の見通しが立たない
+- ✅ 根本原因が特定できない
+- ✅ 同じフレームワークを使い続けるリスク
+- ✅ 新しい技術でより良い設計が可能
+
+これらの理由から、**一から作り直す判断**に至りました。
+
+## ユーザーへの影響
+
+### 改善されたポイント
+
+- **ログインの安定性**: 認証処理がシンプルになり、ログイン問題が解消
+- **読み込み速度**: Next.jsの最適化により、ページ表示が高速化
+- **アクセス性**: ブラウザから直接アクセス可能（アプリダウンロード不要）
+- **更新の即時反映**: 機能追加や修正がすぐにユーザーに届く
+
+### ユーザーが気付く変化
+
+- URLにアクセスするだけで使える
+- ログイン時のエラーや待ち時間が減少
+- 画面遷移がスムーズに
+
+## 失敗から学んだこと
+
+### 1. 早期の本番テスト
+
+開発の後半になってから本番環境でテストするのではなく、**早い段階で実際の環境に近い形で検証する**ことの重要性を学びました。
+
+開発中のテスト環境では問題なく動いていたため、本番環境特有の問題に気付くのが遅れました。もっと早い段階で本番環境に近い条件でテストしていれば、問題を早期に発見し、対処する時間を確保できたはずです。
+
+### 2. 技術選定の視点
+
+フレームワークやツールを選ぶとき、「作りやすさ」だけでなく**「問題が起きたときの対処しやすさ」**も重要な判断基準だと実感しました。
+
+- **情報の多さ**: 困ったときにネットで解決策が見つかるか
+- **コミュニティの活発さ**: 同じ問題に遭遇した人がいるか
+- **デバッグのしやすさ**: エラーの原因を追跡できる仕組みがあるか
+
+どんな技術にも長所・短所がありますが、トラブル発生時にどれだけ対処しやすいかは、プロジェクトの成功に大きく影響します。
+
+### 3. 損切りの判断
+
+修正を続けることと、作り直すことの**判断タイミング**を見極める重要性を学びました。
+
+「もう少し頑張れば直るかも」という思いで修正を続けがちですが、根本原因が不明なまま対症療法を繰り返すと、かえって時間を浪費することがあります。
+
+作り直すという決断は勇気がいりますが、結果的にはより早く、より良いものを作ることにつながりました。
+
+## ユーザーへの感謝
+
+Flutterバージョンで不便をおかけしたユーザーの皆様には、心よりお詫び申し上げます。
+
+この経験を糧に、より安定したサービスを提供できるよう努めていきます。引き続きローストプラスをよろしくお願いいたします。`,
+  tags: ['開発背景'],
+  publishedAt: '2026-01-17',
+  order: 3,
+};
