+++ data/dev-stories/episode-005.ts
@@ -1,0 +1,378 @@
+import type { DevStoryEpisode } from '@/types';
+
+export const EPISODE_005: DevStoryEpisode = {
+  id: 'episode-005',
+  title: 'いろんな資格取得に挑戦中！',
+  subtitle: 'IT・AI資格を楽しく1ヶ月集中で取得',
+  imageUrl: '/dev-stories/qualifications_challenge.png',
+  dialogues: [
+    // 導入
+    {
+      id: 'msg-005-1',
+      characterId: 'asairi',
+      content: 'ねえねえ、最近いろんな資格を取ってるって聞いたんだけど！',
+    },
+    {
+      id: 'msg-005-2',
+      characterId: 'fukairi',
+      content: 'うん、IT系とAI系の資格に挑戦してるんだ。\n実は生成AIパスポートとITパスポートに合格したんだよ。',
+    },
+    {
+      id: 'msg-005-3',
+      characterId: 'asairi',
+      content: 'すごい！おめでとう！他にも受けてるの？',
+    },
+    {
+      id: 'msg-005-4',
+      characterId: 'fukairi',
+      content: 'G検定も受けて、今は合格発表待ちなんだ。\nあと2月22日に基本情報技術者試験を受ける予定だよ。',
+    },
+    // 生成AIパスポート
+    {
+      id: 'msg-005-5',
+      characterId: 'asairi',
+      content: '生成AIパスポートって何？どんな資格なの？',
+    },
+    {
+      id: 'msg-005-6',
+      characterId: 'fukairi',
+      content: '生成AIを活用するための知識を証明する資格なんだ。\nChatGPTとかの仕組みや、ビジネス活用、著作権の問題なんかを学べるよ。',
+    },
+    {
+      id: 'msg-005-7',
+      characterId: 'asairi',
+      content: '今まさに必要な知識だね！どうやって勉強したの？',
+    },
+    {
+      id: 'msg-005-8',
+      characterId: 'fukairi',
+      content: '「生成AIパスポートテキスト&問題集」と「生成AIパスポート、トレ速」っていうアプリを使ったよ。\n普段からAIを使ってるから、実感を持って勉強できたんだ。',
+    },
+    {
+      id: 'msg-005-9',
+      characterId: 'asairi',
+      content: 'ローストプラスの開発でAIを使ってるもんね！',
+    },
+    {
+      id: 'msg-005-10',
+      characterId: 'fukairi',
+      content: 'そうなんだ。実際に使っている経験があると、テキストの内容がスッと入ってくるんだよね。',
+    },
+    // ITパスポート
+    {
+      id: 'msg-005-11',
+      characterId: 'asairi',
+      content: 'ITパスポートはどんな資格？',
+    },
+    {
+      id: 'msg-005-12',
+      characterId: 'fukairi',
+      content: 'ITの基礎知識を幅広く学べる国家資格だよ。\n経営戦略、マネジメント、テクノロジーの3分野から出題されるんだ。',
+    },
+    {
+      id: 'msg-005-13',
+      characterId: 'asairi',
+      content: '国家資格なんだ！難しそう...',
+    },
+    {
+      id: 'msg-005-14',
+      characterId: 'fukairi',
+      content: '実は思ったより取り組みやすかったよ。\n「過去問道場」っていう無料のサイトと、アプリ、あとAnkiっていう分散学習ツールを使ったんだ。',
+    },
+    {
+      id: 'msg-005-15',
+      characterId: 'asairi',
+      content: 'Anki？聞いたことないな。どんなツールなの？',
+    },
+    {
+      id: 'msg-005-16',
+      characterId: 'fukairi',
+      content: 'Ankiは忘れる頃に問題を出してくれるから、効率よく記憶に定着するんだ。\nスマホでどこでも勉強できるから、通勤中とか隙間時間に活用できたんだよ。',
+    },
+    // G検定
+    {
+      id: 'msg-005-17',
+      characterId: 'asairi',
+      content: 'G検定は？AIの資格だよね？',
+    },
+    {
+      id: 'msg-005-18',
+      characterId: 'fukairi',
+      content: 'そう、日本ディープラーニング協会が主催するAI・機械学習の検定だよ。\nディープラーニングの仕組みやビジネス活用について問われるんだ。',
+    },
+    {
+      id: 'msg-005-19',
+      characterId: 'asairi',
+      content: '難しそう...どんな勉強をしたの？',
+    },
+    {
+      id: 'msg-005-20',
+      characterId: 'fukairi',
+      content: '「ディープラーニングG検定（ジェネラリスト）最強の合格テキスト」と、AI問題集アプリを使ったよ。\n数学的な部分は難しかったけど、概念を理解することに集中したんだ。',
+    },
+    {
+      id: 'msg-005-21',
+      characterId: 'asairi',
+      content: '試験はどうだった？',
+    },
+    {
+      id: 'msg-005-22',
+      characterId: 'fukairi',
+      content: 'オンライン試験で、100分で145問を解くんだ。\n時間との戦いだったけど、なんとか最後まで解けたよ。',
+    },
+    {
+      id: 'msg-005-23',
+      characterId: 'asairi',
+      content: '145問！？それはすごいね！結果が楽しみだね。',
+    },
+    // 基本情報技術者試験
+    {
+      id: 'msg-005-24',
+      characterId: 'asairi',
+      content: '次は基本情報技術者試験を受けるんだよね？',
+    },
+    {
+      id: 'msg-005-25',
+      characterId: 'fukairi',
+      content: 'うん、2月22日に受験予定なんだ。\nITエンジニアの登竜門と言われる国家資格だよ。',
+    },
+    {
+      id: 'msg-005-26',
+      characterId: 'asairi',
+      content: 'ITパスポートより難しいの？',
+    },
+    {
+      id: 'msg-005-27',
+      characterId: 'fukairi',
+      content: 'うん、かなりレベルアップするよ。\nアルゴリズムやプログラミング、ネットワークなど、より専門的な内容が出題されるんだ。',
+    },
+    {
+      id: 'msg-005-28',
+      characterId: 'asairi',
+      content: '今どんな勉強をしてるの？',
+    },
+    {
+      id: 'msg-005-29',
+      characterId: 'fukairi',
+      content: 'かやのき先生の参考書で基礎を固めて、過去問道場で問題演習をしてるよ。\n科目A・科目Bの両方をバランスよく対策中なんだ。',
+    },
+    {
+      id: 'msg-005-30',
+      characterId: 'asairi',
+      content: '応援してるね！頑張って！',
+    },
+    // 勉強のコツ
+    {
+      id: 'msg-005-31',
+      characterId: 'asairi',
+      content: 'こんなにたくさんの資格を取るコツってあるの？',
+    },
+    {
+      id: 'msg-005-32',
+      characterId: 'fukairi',
+      content: '一番大事なのは「楽しむこと」かな。\n新しいことを学ぶのって、本来ワクワクするものだと思うんだ。',
+    },
+    {
+      id: 'msg-005-33',
+      characterId: 'asairi',
+      content: '勉強が楽しいってすごいね！',
+    },
+    {
+      id: 'msg-005-34',
+      characterId: 'fukairi',
+      content: 'あとは「短期集中」がポイントかな。\n長期間ダラダラやるより、約1ヶ月でギュッと集中した方が効率がいいんだ。',
+    },
+    {
+      id: 'msg-005-35',
+      characterId: 'asairi',
+      content: '確かに、モチベーションも維持しやすいよね！',
+    },
+    {
+      id: 'msg-005-36',
+      characterId: 'fukairi',
+      content: 'それと、スマホアプリや過去問サイトを活用すると、スキマ時間を有効に使えるよ。\n電車の中とか、ちょっとした待ち時間でも勉強できるんだ。',
+    },
+    // まとめ
+    {
+      id: 'msg-005-37',
+      characterId: 'asairi',
+      content: '資格を取ることで何か変わった？',
+    },
+    {
+      id: 'msg-005-38',
+      characterId: 'fukairi',
+      content: '知識が体系的に整理されたのが大きいかな。\nなんとなく使ってたAIやITの仕組みを、ちゃんと理解できるようになったんだ。',
+    },
+    {
+      id: 'msg-005-39',
+      characterId: 'asairi',
+      content: 'ローストプラスの開発にも活かせそう？',
+    },
+    {
+      id: 'msg-005-40',
+      characterId: 'fukairi',
+      content: 'もちろん！セキュリティやデータベースの知識とか、実際の開発に役立ってるよ。\n資格の勉強は「使える知識」を身につけるのにとても良い方法だと思うんだ。',
+    },
+    {
+      id: 'msg-005-41',
+      characterId: 'asairi',
+      content: 'これから資格を取ろうとしている人へメッセージはある？',
+    },
+    {
+      id: 'msg-005-42',
+      characterId: 'fukairi',
+      content: '完璧を目指さなくていいから、まず始めてみることが大事だよ。\n楽しみながら、自分のペースで挑戦してほしいな。一緒に頑張ろう！',
+    },
+  ],
+  detailContent: `IT・AI関連の資格取得に挑戦した体験談をお届けします。楽しく効率的に資格を取得するコツも紹介します！
+
+## 取得・挑戦中の資格一覧
+
+| 資格名 | 状況 | 試験形式 |
+|--------|------|----------|
+| **生成AIパスポート** | ✅ 合格 | オンライン |
+| **ITパスポート** | ✅ 合格 | CBT（テストセンター） |
+| **G検定** | ⏳ 合格発表待ち | オンライン |
+| **基本情報技術者試験** | 📚 勉強中（2月22日受験予定） | CBT（テストセンター） |
+
+## 生成AIパスポート
+
+### 資格の概要
+
+生成AIパスポートは、ChatGPTなどの生成AIを活用するための知識を証明する資格です。
+
+**出題範囲：**
+- 生成AIの仕組み（LLM、プロンプト、トランスフォーマーなど）
+- ビジネス活用と事例
+- 著作権・知的財産権の問題
+- AIの倫理とリスク
+- 最新のAI動向
+
+### 勉強方法
+
+- **「生成AIパスポートテキスト&問題集」**：基礎知識を体系的に学習
+- **「生成AIパスポート、トレ速」アプリ**：スマホで手軽に演習
+- **実践経験**：日常的にAIツールを使う
+
+### 合格のポイント
+
+普段からChatGPTやClaudeを使っていたので、実感を持って勉強できました。テキストの内容が「あ、これ知ってる！」と繋がる瞬間が多く、楽しく学べました。
+
+## ITパスポート
+
+### 資格の概要
+
+ITパスポートは、ITの基礎知識を幅広く問う国家資格です。
+
+**出題分野：**
+- **ストラテジ系**（経営戦略、マーケティング、法務など）
+- **マネジメント系**（プロジェクト管理、サービスマネジメントなど）
+- **テクノロジ系**（ネットワーク、セキュリティ、データベースなど）
+
+### 勉強方法
+
+- **過去問道場**（https://www.itpassportsiken.com/）：無料で過去問が解ける神サイト
+- **スマホアプリ**：通勤中や隙間時間に活用
+- **Anki**：分散学習ツール。忘れる頃に問題を出してくれるので効率よく記憶に定着する
+- **解説を読み込む**：間違えた問題は必ず理解する
+
+### 合格のポイント
+
+過去問道場とAnkiを徹底的に活用しました。電車の中、待ち時間、寝る前など、スマホがあればどこでも勉強できるのが良かったです。
+
+## G検定（ジェネラリスト検定）
+
+### 資格の概要
+
+G検定は、日本ディープラーニング協会（JDLA）が主催するAI・機械学習の検定です。
+
+**出題範囲：**
+- 人工知能の歴史と定義
+- 機械学習の手法（教師あり学習、教師なし学習など）
+- ディープラーニングの仕組み
+- CNNやRNNなどのニューラルネットワーク
+- AI活用の法律・倫理
+
+### 勉強方法
+
+- **「ディープラーニングG検定（ジェネラリスト）最強の合格テキスト」**：基礎をしっかり学習
+- **AI問題集アプリ**：スマホで手軽に演習
+- **過去問対策**：時間配分を意識して解く
+
+### 試験の感想
+
+100分で145問という問題数。最初は焦りましたが、サクサク解くことを意識して最後まで完走できました。合格発表が楽しみです！
+
+## 基本情報技術者試験（勉強中）
+
+### 資格の概要
+
+基本情報技術者試験は、ITエンジニアの登竜門と言われる国家資格です。
+
+**試験構成：**
+- **科目A**：60問（90分）- 知識問題
+- **科目B**：20問（100分）- アルゴリズム・情報セキュリティ
+
+### 勉強方法
+
+- **かやのき式 基本情報技術者**：わかりやすい解説で基礎を固める
+- **過去問道場**：科目A対策に最適
+- **アルゴリズム対策**：トレース練習を重点的に
+
+### 現在の進捗
+
+2月22日の受験に向けて追い込み中です。科目Aは過去問で8割以上取れるようになってきたので、科目Bのアルゴリズム対策に力を入れています。
+
+## 1ヶ月集中で資格を取るコツ
+
+### 1. 楽しむこと
+
+勉強を「やらなきゃいけないこと」ではなく「新しいことを知る楽しみ」と捉えると、モチベーションが続きます。
+
+### 2. 短期集中
+
+長期間ダラダラ勉強するより、約1ヶ月でギュッと集中した方が効率的です。
+
+- **ITパスポート**：約1ヶ月
+- **生成AIパスポート**：約1ヶ月
+- **G検定**：約1ヶ月
+
+### 3. スキマ時間の活用
+
+- 通勤時間
+- 待ち時間
+- 寝る前の10分
+
+スマホアプリや過去問サイトを使えば、どこでも勉強できます。
+
+### 4. 完璧を目指さない
+
+100%理解してから次に進むのではなく、まず全体像をつかんでから深掘りする方が効率的です。
+
+### 5. 実践と結びつける
+
+資格の勉強で得た知識を、実際のプロジェクトで使うと定着します。ローストプラスの開発でも、セキュリティやデータベースの知識が役立っています。
+
+## おすすめの勉強ツール
+
+| ツール | 対象資格 | 特徴 |
+|--------|----------|------|
+| **過去問道場** | ITパスポート、基本情報 | 無料で過去問が解ける |
+| **AI問題集アプリ** | G検定 | スマホで手軽に演習 |
+| **Anki** | 全般 | 分散学習で効率よく記憶に定着 |
+| **公式テキスト** | 各資格 | 体系的に学べる |
+| **YouTube** | 全般 | 解説動画で理解を深める |
+
+## これから資格を取る人へ
+
+資格取得は、知識を体系的に整理する良い機会です。
+
+- **始めることが大事**：完璧な準備を待たずに、まず始めてみる
+- **楽しむこと**：新しいことを学ぶワクワク感を大切に
+- **自分のペースで**：他人と比べず、自分に合った勉強法を見つける
+
+一緒に頑張りましょう！`,
+  tags: ['資格', 'IT', '勉強法', '体験談'],
+  publishedAt: '2026-01-24',
+  order: 5,
+};
