+++ data/dev-stories/episode-006.ts
@@ -1,0 +1,355 @@
+import type { DevStoryEpisode } from '@/types';
+
+export const EPISODE_006: DevStoryEpisode = {
+  id: 'episode-006',
+  title: 'Claude Codeを使ってみた',
+  subtitle: '文字でAIと会話しながら開発できる新しいツール',
+  imageUrl: '/dev-stories/claude_code.png',
+  dialogues: [
+    // 導入（5メッセージ）
+    {
+      id: 'msg-006-1',
+      characterId: 'press',
+      content: '最近「Claude Code」っていうツールを使い始めたって聞いたけど、どういうものなの？',
+    },
+    {
+      id: 'msg-006-2',
+      characterId: 'siphon',
+      content: '簡単に言うと、文字で会話しながらアプリを作れるAIアシスタントなんだ。',
+    },
+    {
+      id: 'msg-006-3',
+      characterId: 'press',
+      content: 'Cursorとは違うの？',
+    },
+    {
+      id: 'msg-006-4',
+      characterId: 'siphon',
+      content: 'Cursorはエディタ画面の中でAIと対話するスタイル。Claude Codeはターミナルっていう黒い画面でテキストだけでやり取りするんだ。シンプルだけど、その分パワフルなんだよね。',
+    },
+    {
+      id: 'msg-006-5',
+      characterId: 'press',
+      content: 'なんだかレトロな感じがするけど、どうしてそっちを使うの？',
+    },
+    // AIアシスタントの種類（8メッセージ）
+    {
+      id: 'msg-006-6',
+      characterId: 'siphon',
+      content: '実はアプリ開発を手伝ってくれるAIにも色々なタイプがあるんだ。',
+    },
+    {
+      id: 'msg-006-7',
+      characterId: 'press',
+      content: 'どんな種類があるの？',
+    },
+    {
+      id: 'msg-006-8',
+      characterId: 'siphon',
+      content: '一つ目は「エディタ統合型」。CursorやAntigravityみたいに、コードエディタの中にAIが組み込まれているタイプだね。',
+    },
+    {
+      id: 'msg-006-9',
+      characterId: 'press',
+      content: 'それは前から使ってたよね。他には？',
+    },
+    {
+      id: 'msg-006-10',
+      characterId: 'siphon',
+      content: '二つ目が「ターミナル型」。Claude Codeがこれにあたるね。テキストだけでやり取りするから、動作が軽くて柔軟性が高いんだ。',
+    },
+    {
+      id: 'msg-006-11',
+      characterId: 'press',
+      content: 'コードを書かなくてもいいタイプもあるって聞いたけど？',
+    },
+    {
+      id: 'msg-006-12',
+      characterId: 'siphon',
+      content: 'LovableやBoltみたいに「こんなアプリが欲しい」って伝えるだけで作ってくれるツールもあるよ。ただ、細かい調整が難しいんだよね。',
+    },
+    {
+      id: 'msg-006-13',
+      characterId: 'press',
+      content: 'それぞれ得意なことが違うんだね。Claude Codeは何が得意なの？',
+    },
+    // Claude Codeの便利機能（12メッセージ）
+    {
+      id: 'msg-006-14',
+      characterId: 'siphon',
+      content: 'Claude Codeのすごいところは「自分で考えて動いてくれる」ことなんだ。例えば「この機能を直して」って言うと…',
+    },
+    {
+      id: 'msg-006-15',
+      characterId: 'press',
+      content: 'どうなるの？',
+    },
+    {
+      id: 'msg-006-16',
+      characterId: 'siphon',
+      content: 'まず自分でファイルを探して、中身を読んで、問題を見つけて、直すところまで全部やってくれるんだ。まるで優秀な助手がいるみたい。',
+    },
+    {
+      id: 'msg-006-17',
+      characterId: 'press',
+      content: 'それはすごい！他にも便利な機能はある？',
+    },
+    {
+      id: 'msg-006-18',
+      characterId: 'siphon',
+      content: '「スキル」っていう機能があって、よく使う作業を一言で頼めるんだ。例えば「/commit」って打つだけで、変更を記録してくれる。',
+    },
+    {
+      id: 'msg-006-19',
+      characterId: 'press',
+      content: '料理のレシピみたいに、定番の手順が登録されてるってこと？',
+    },
+    {
+      id: 'msg-006-20',
+      characterId: 'siphon',
+      content: 'そうそう！自分だけのオリジナルレシピも作れるんだよ。ローストプラスでは「UIデザインを確認する」スキルとか作ってるんだ。',
+    },
+    {
+      id: 'msg-006-21',
+      characterId: 'press',
+      content: '他のツールとも連携できるって聞いたけど？',
+    },
+    {
+      id: 'msg-006-22',
+      characterId: 'siphon',
+      content: 'うん、「外部ツール連携」っていう仕組みがあってね。ローストプラスの開発では3つ使ってるんだ。',
+    },
+    {
+      id: 'msg-006-23',
+      characterId: 'press',
+      content: 'どんなことができるの？',
+    },
+    {
+      id: 'msg-006-24',
+      characterId: 'siphon',
+      content: '一つ目はコードの中身を賢く探せる機能。二つ目は最新のマニュアルを自動で調べてくれる機能。三つ目はブラウザ画面を見て確認してくれる機能。',
+    },
+    {
+      id: 'msg-006-25',
+      characterId: 'press',
+      content: 'AIがブラウザ画面を見れるの？',
+    },
+    {
+      id: 'msg-006-26',
+      characterId: 'siphon',
+      content: 'そう！「このボタンをクリックして」って頼むと、実際に画面を見ながら操作してくれるんだ。表示がおかしくないかも確認できるよ。',
+    },
+    // 開発効率の向上（6メッセージ）
+    {
+      id: 'msg-006-27',
+      characterId: 'press',
+      content: 'ローストプラスの開発でどう役立ってるの？',
+    },
+    {
+      id: 'msg-006-28',
+      characterId: 'siphon',
+      content: '実はこの「開発秘話」ページ自体がClaude Codeで作られてるんだ。新しいエピソードを追加するのも、ほとんど自動でやってくれる。',
+    },
+    {
+      id: 'msg-006-29',
+      characterId: 'press',
+      content: 'へぇ！具体的にどう楽になったの？',
+    },
+    {
+      id: 'msg-006-30',
+      characterId: 'siphon',
+      content: '例えば「焙煎タイマーの表示を大きくして」って頼むと、関連するファイルを探して、コードを直して、画面で確認するところまでやってくれるんだ。',
+    },
+    {
+      id: 'msg-006-31',
+      characterId: 'press',
+      content: 'コーヒーを淹れてる間に作業が進んでそうだね。',
+    },
+    {
+      id: 'msg-006-32',
+      characterId: 'siphon',
+      content: 'まさにそんな感じ！以前は自分で全部やってたから、体感で3倍くらい速くなったかな。',
+    },
+    // Cursorとの比較（4メッセージ）
+    {
+      id: 'msg-006-33',
+      characterId: 'press',
+      content: 'じゃあCursorはもう使わないの？',
+    },
+    {
+      id: 'msg-006-34',
+      characterId: 'siphon',
+      content: 'いや、両方使い分けてるよ。Cursorは画面を見ながら細かく調整したい時に便利。Claude Codeは「この機能を作って」って大きく任せたい時に使うんだ。',
+    },
+    {
+      id: 'msg-006-35',
+      characterId: 'press',
+      content: 'コーヒーで例えると、どんな感じ？',
+    },
+    {
+      id: 'msg-006-36',
+      characterId: 'siphon',
+      content: 'Cursorはハンドドリップみたいに丁寧に一杯ずつ淹れる感じ。Claude Codeは優秀なバリスタに「美味しいコーヒーをお願い」って任せる感じかな。',
+    },
+    // まとめ（4メッセージ）
+    {
+      id: 'msg-006-37',
+      characterId: 'press',
+      content: 'AIと一緒に開発する時代が本格的に来たんだね。',
+    },
+    {
+      id: 'msg-006-38',
+      characterId: 'siphon',
+      content: 'そうだね。でも大事なのは「何を作りたいか」っていうアイデア。道具が便利になっても、そこは変わらないと思うんだ。',
+    },
+    {
+      id: 'msg-006-39',
+      characterId: 'press',
+      content: 'ローストプラスも、コーヒーへの愛情があってこそってことだね。',
+    },
+    {
+      id: 'msg-006-40',
+      characterId: 'siphon',
+      content: 'その通り！これからもAIの力を借りながら、もっと使いやすいアプリにしていきたいな。',
+    },
+  ],
+  detailContent: `Claude Codeは、文字で会話しながらアプリを作れるAIアシスタントです。この記事では、AIアシスタントの種類からClaude Codeの便利な機能、そしてローストプラスでどう活用しているかをわかりやすく紹介します。
+
+## 開発を手伝ってくれるAIアシスタントとは
+
+最近、アプリ開発を手伝ってくれるAIツールが増えています。「こんな機能を作って」とお願いすると、自分でファイルを探して、コードを書いて、動作確認まで手伝ってくれます。まるで優秀な助手がそばにいるような感覚です。
+
+## AIアシスタントの種類
+
+開発を手伝ってくれるAIツールには、いくつかのタイプがあります。
+
+| タイプ | 特徴 | 代表的なツール |
+|--------|------|--------------|
+| **エディタ統合型** | コードエディタにAIが組み込まれている | Cursor, Antigravity |
+| **ターミナル型** | テキストベースで柔軟に対話 | Claude Code, Codex |
+| **ノーコード型** | 自然言語だけでアプリを生成 | Lovable, Bolt |
+
+### エディタ統合型
+
+CursorやAntigravityは、コードエディタの中にAIが組み込まれています。
+
+- **良いところ**: 視覚的で分かりやすい、コードのハイライトや補完が充実
+- **苦手なところ**: エディタに依存する
+
+### ターミナル型
+
+Claude Codeは、ターミナル（黒い画面）でテキストベースでやり取りします。
+
+- **良いところ**: 動作が軽い、スクリプトとの連携がしやすい、拡張性が高い
+- **苦手なところ**: 最初は慣れが必要
+
+### ノーコード型
+
+LovableやBoltは、「こんなアプリが欲しい」と自然言語で伝えるだけで形にしてくれます。
+
+- **良いところ**: プログラミング経験がなくても始められる
+- **苦手なところ**: 細かいカスタマイズが難しい
+
+## Claude Codeの便利な機能
+
+### 「自分で考えて動いてくれる」
+
+Claude Codeの最大の特徴は、自律的に作業してくれること。「この機能を直して」とお願いすると：
+
+1. 関連するファイルを自分で探す
+2. 中身を読んで問題を見つける
+3. コードを直す
+4. 動作確認まで手伝ってくれる
+
+まるで優秀なアシスタントがそばにいるような体験ができます。
+
+### スキル機能 - 定番の作業を一言で
+
+よく使う作業を「レシピ」のように登録しておいて、一言で呼び出せる機能です。
+
+**最初から使えるもの：**
+- \`/commit\` - 変更を記録
+- \`/review-pr\` - 変更内容をチェック
+
+**ローストプラスで作ったオリジナル：**
+- \`/roastplus-ui\` - デザインのルールを確認
+- \`/nextjs-firestore\` - データベース連携の書き方を確認
+
+料理でいえば、「チャーハンを作って」と言うだけで、材料を出して、切って、炒めて、盛り付けまでやってくれるイメージです。
+
+### 外部ツール連携 - AIの能力を拡張
+
+Claude Codeは他のツールと連携して、もっと賢くなれます。ローストプラスでは3つの連携を使っています：
+
+| 連携ツール | 何ができる？ |
+|-----------|------------|
+| **Serena** | コードの中身を賢く探す。「この機能を使っている場所を全部見つけて」が可能 |
+| **Context7** | 最新のマニュアルを自動で調べる。「この書き方で合ってる？」を確認 |
+| **Chrome DevTools** | ブラウザ画面を見て確認。「このボタンをクリックして」が可能 |
+
+### おまかせモード - 信頼できる操作は自動で
+
+通常は「この変更をしていい？」と毎回確認されますが、信頼できる操作は自動でOKするように設定できます。
+
+**安心ポイント：**
+- 変更履歴は全部残るので、いつでも元に戻せる
+- 最初は慎重に、慣れてから範囲を広げる
+
+## ローストプラスでの活用例
+
+### この「開発秘話」ページを作った
+
+実はこの開発秘話ページ自体が、Claude Codeで作られています。
+
+**作業の流れ：**
+1. 「新しいエピソードを追加したい」とお願い
+2. Claude Codeがファイル構造を理解
+3. 新しいエピソードのコードを生成
+4. 画面で表示を確認
+5. 完成！
+
+### 機能の改善
+
+「焙煎タイマーの表示を大きくして」とお願いすると：
+- 関連するファイルを探す
+- デザインのコードを修正
+- 画面で確認して完了
+
+コーヒーを淹れている間に作業が進んでいることも。
+
+### デザインの確認
+
+ブラウザ連携を使って、実際の画面を見ながら開発できます。
+- 「この画面のスクリーンショットを撮って」
+- 「このボタンをクリックして動作確認して」
+- 「エラーが出てないか確認して」
+
+## CursorとClaude Codeの使い分け
+
+どちらも素晴らしいツールなので、場面で使い分けています。
+
+| 場面 | 使うツール | 理由 |
+|------|-----------|------|
+| 細かい調整 | Cursor | 画面を見ながら微調整しやすい |
+| 大きな機能追加 | Claude Code | 「作って」と任せられる |
+| デザイン確認 | Claude Code | ブラウザ連携が便利 |
+| 学習中 | Cursor | 視覚的で分かりやすい |
+
+**コーヒーで例えると：**
+- **Cursor**: ハンドドリップ。丁寧に一杯ずつ淹れる
+- **Claude Code**: 優秀なバリスタ。「美味しいコーヒーをお願い」と任せられる
+
+## まとめ
+
+Claude Codeは、文字で会話するシンプルなスタイルながら、自分で考えて動いてくれる優秀なAIアシスタントです。
+
+- **自律的**: ファイルを探して、読んで、直すまで自分でやってくれる
+- **レシピ機能**: よく使う作業を一言で頼める
+- **連携機能**: 他のツールと組み合わせてもっと賢く
+- **安心設計**: 変更履歴が残るので、いつでも元に戻せる
+
+AIと一緒に開発する時代が来ています。でも大切なのは、「何を作りたいか」というアイデアと情熱。ローストプラスも、コーヒーへの愛情があってこそ。これからもAIの力を借りながら、もっと使いやすいアプリを目指していきます。`,
+  tags: ['Claude Code', 'AI開発', 'ツール紹介', '開発効率'],
+  publishedAt: '2026-01-25',
+  order: 6,
+};
