+++ data/dev-stories/episodes.ts
@@ -1,1700 +1,20 @@
 import type { DevStoryEpisode } from '@/types';
 
+import { EPISODE_001 } from './episode-001';
+import { EPISODE_002 } from './episode-002';
+import { EPISODE_003 } from './episode-003';
+import { EPISODE_004 } from './episode-004';
+import { EPISODE_005 } from './episode-005';
+import { EPISODE_006 } from './episode-006';
+
 // 開発秘話エピソード
 export const DEV_STORY_EPISODES: DevStoryEpisode[] = [
-  {
-    id: 'episode-001',
-    title: 'ローストプラス誕生秘話',
-    subtitle: 'アプリ開発のきっかけ',
-    imageUrl: '/dev-stories/roast_plus_origin.png',
-    dialogues: [
-      {
-        id: 'msg-001-1',
-        characterId: 'asairi',
-        content: 'ねえねえ、ローストプラスってどうして作ろうと思ったの？',
-      },
-      {
-        id: 'msg-001-2',
-        characterId: 'fukairi',
-        content: '実は担当表のマンネリ化が課題だったんだ。\n掃除とローストの担当を6人で回してたんだけどね。',
-      },
-      {
-        id: 'msg-001-3',
-        characterId: 'asairi',
-        content: 'マンネリ化？どういうこと？',
-      },
-      {
-        id: 'msg-001-4',
-        characterId: 'fukairi',
-        content: '2列3行の担当表があって、左の列は上に、右の列は下にずらすっていうルールでシャッフルしてたんだ。',
-      },
-      {
-        id: 'msg-001-5',
-        characterId: 'asairi',
-        content: 'あー、それだと同じペアになっちゃうってこと？',
-      },
-      {
-        id: 'msg-001-6',
-        characterId: 'fukairi',
-        content: 'そうそう！3日経つと元の組み合わせに戻っちゃうんだよね。\n周期的なパターンになって、特定の人に偏りがちだった。',
-      },
-      {
-        id: 'msg-001-7',
-        characterId: 'asairi',
-        content: 'なるほど〜！それでランダムにシャッフルできる機能が欲しくなったんだね！',
-      },
-      {
-        id: 'msg-001-8',
-        characterId: 'fukairi',
-        content: 'うん。ランダムだけじゃなくて、公平性も大事でね。\n誰かに偏らず、みんなが平等に担当できるアプリを作りたかったんだ。',
-      },
-    ],
-    detailContent: `ローストプラスは、担当表運用の課題を解決するために生まれました。
-
-## 従来の課題
-
-6人で掃除・ローストの担当を回していましたが、以下の問題がありました：
-
-- **固定的なシャッフルルール**: 2列3行の担当表を「左は上に、右は下に」というルールでシャッフルしていた
-- **周期問題**: 3日後には同じペアの組み合わせに戻ってしまう
-- **偏りの発生**: 特定の人が同じ担当になりやすかった
-
-## ローストプラスで解決
-
-これらの課題を解決するため、ローストプラスでは以下の機能を実装しました：
-
-- **ランダムシャッフル**: 予測できないランダムな組み合わせを生成
-- **公平性の担保**: 誰かに偏ることなく、平等に担当を割り振る仕組み
-
-マンネリ化した担当表を解決するアプリとして開発がスタートしました。`,
-    tags: ['開発背景', 'コンセプト', '担当表'],
-    publishedAt: '2026-01-16',
-    order: 1,
-  },
-  {
-    id: 'episode-002',
-    title: 'バイブコーディングとは？',
-    subtitle: 'AIと一緒にアプリを作る新しい方法',
-    imageUrl: '/dev-stories/vibe_coding.png',
-    dialogues: [
-      {
-        id: 'msg-002-1',
-        characterId: 'dori',
-        content: '「バイブコーディング」っていう言葉を聞いたんだけど、どんなものなのかな？',
-      },
-      {
-        id: 'msg-002-2',
-        characterId: 'server',
-        content: 'それはね、AIと会話しながらアプリを作っていく新しい方法だよ。',
-      },
-      {
-        id: 'msg-002-3',
-        characterId: 'dori',
-        content: '会話でアプリが作れるなんて、なんだか不思議だな...どういう仕組みなのかな？',
-      },
-      {
-        id: 'msg-002-4',
-        characterId: 'server',
-        content: '「こんな機能が欲しいな」って言葉で伝えると、AIがプログラムのコードを書いてくれるんだ。',
-      },
-      {
-        id: 'msg-002-5',
-        characterId: 'dori',
-        content: 'ということは...プログラミング言語を覚えなくても大丈夫なのかな？',
-      },
-      {
-        id: 'msg-002-6',
-        characterId: 'server',
-        content: 'そうだね。だから「バイブ」＝雰囲気やイメージで伝えるっていう意味なんだよ。',
-      },
-      {
-        id: 'msg-002-7',
-        characterId: 'dori',
-        content: 'そうなんだ...このローストプラスもそうやって作られたのかな？',
-      },
-      {
-        id: 'msg-002-8',
-        characterId: 'server',
-        content: '実はね、ローストプラスは、ほぼ全部バイブコーディングで作られたんだよ。',
-      },
-      {
-        id: 'msg-002-9',
-        characterId: 'dori',
-        content: 'このアプリ全部が...！それはすごいことだね。',
-      },
-      {
-        id: 'msg-002-10',
-        characterId: 'server',
-        content: '「担当表をシャッフルする機能が欲しい」みたいに伝えるだけで、形になっていくんだ。',
-      },
-      {
-        id: 'msg-002-11',
-        characterId: 'dori',
-        content: 'どんなAIツールを使っているのかな？',
-      },
-      {
-        id: 'msg-002-12',
-        characterId: 'server',
-        content: 'メインはCursorっていうツールで、他にもCodex、Antigravity、最近はClaude Codeにも挑戦中なんだ。',
-      },
-      {
-        id: 'msg-002-13',
-        characterId: 'dori',
-        content: 'そんなにたくさんあるんだ...使い分けているのかな？',
-      },
-      {
-        id: 'msg-002-14',
-        characterId: 'server',
-        content: 'うん。ツールによって得意なことが違うから、場面に合わせて選んでいるよ。',
-      },
-      {
-        id: 'msg-002-15',
-        characterId: 'dori',
-        content: 'とても便利そうだけど...難しいこともあるのかな？',
-      },
-      {
-        id: 'msg-002-16',
-        characterId: 'server',
-        content: 'あるね。AIに意図を正確に伝えるのが難しいし、思った通りにならないことも多いんだ。\n何度もやり直す根気が必要かもね。',
-      },
-      {
-        id: 'msg-002-17',
-        characterId: 'server',
-        content: 'あと、AIが書いたコードを理解せずに使うと、セキュリティの問題が起きることもあるよ。\nちゃんと確認することが大切だね。',
-      },
-      {
-        id: 'msg-002-18',
-        characterId: 'dori',
-        content: '便利だけど、ちゃんと向き合う姿勢が大事なんだね。',
-      },
-      {
-        id: 'msg-002-19',
-        characterId: 'server',
-        content: 'そうだね。でも挑戦する価値は十分あると思うよ。\nアイデアを形にする敷居がぐっと下がったからね。',
-      },
-    ],
-    detailContent: `バイブコーディングは、AIとの対話でアプリを作る新しいプログラミング手法です。
-
-## バイブコーディングとは
-
-**バイブコーディング**とは、プログラミング言語を使わず、自然な言葉でAIに「こんな機能が欲しい」と伝えてコードを生成してもらう開発方法です。
-
-「バイブ（Vibe）」は英語で「雰囲気」や「感覚」という意味。細かい技術的な指示ではなく、**やりたいことのイメージを伝えるだけ**でアプリが作れることから、この名前がつきました。
-
-## 従来のプログラミングとの違い
-
-| 従来のプログラミング | バイブコーディング |
-|---------------------|-------------------|
-| プログラミング言語を習得する必要がある | 自然な言葉で伝えるだけ |
-| コードを1行ずつ書く | AIがコードを生成 |
-| エラーを自分で解決 | AIと一緒に解決 |
-| 学習に長い時間がかかる | すぐに始められる |
-
-## メリット
-
-- **プログラミング経験がなくてもアプリが作れる**
-- **アイデアを素早く形にできる**
-- **試行錯誤のスピードが格段に上がる**
-
-## ローストプラスで使っているAIツール
-
-| ツール | 特徴 | このアプリでの使い方 |
-|--------|------|---------------------|
-| **Cursor** | AIエディタのパイオニア | メインの開発ツールとして日常的に使用 |
-| **Codex** | OpenAI製・高い推論力 | 複雑なバックエンド機能の修正 |
-| **Antigravity** | 最新のAI系IDE | フロントエンド改善、シャッフルロジック改善 |
-| **Claude Code** | ターミナルベースの対話型 | 最近挑戦中の新しいツール |
-
-それぞれのツールには得意分野があり、場面に合わせて使い分けています。
-
-## デメリット・注意点
-
-バイブコーディングは便利ですが、難しさや注意点もあります：
-
-- **適切な指示を出す難しさ**: AIに意図を正確に伝えるスキルが必要
-- **根気が必要**: 思い通りにならないことも多く、何度もやり直すことがある
-- **評価力が必要**: AIの提案が正しいかどうか判断する力が求められる
-- **セキュリティリスク**: 理解せずにコードを採用すると、危険な場合がある
-
-**大切なのは、AIに任せきりにせず、一緒に作っていく姿勢です。**
-
-## このアプリでの実例
-
-ローストプラスでは、以下の機能がバイブコーディングで作られました：
-
-- 担当表シャッフル機能
-- 試飲感想記録機能
-- この「開発秘話」ページ自体も！
-
-非エンジニアでも、アイデアがあれば形にできる時代になりました。`,
-    tags: ['バイブコーディング', 'AI', '開発手法'],
-    publishedAt: '2026-01-16',
-    order: 2,
-  },
-  {
-    id: 'episode-003',
-    title: 'ローストプラスは一度作り直している',
-    subtitle: 'FlutterからNext.jsへの技術移行',
-    imageUrl: '/dev-stories/flutter_to_nextjs.png',
-    dialogues: [
-      {
-        id: 'msg-003-1',
-        characterId: 'kettle',
-        content: 'ローストプラスって、最初はFlutterで作ってたんだろ？',
-      },
-      {
-        id: 'msg-003-2',
-        characterId: 'mill',
-        content: 'そうなんだ。Flutterっていうのは、1つのコードでiPhoneとAndroid両方のアプリが作れるフレームワーク（開発ツール）なんだよね。',
-      },
-      {
-        id: 'msg-003-3',
-        characterId: 'kettle',
-        content: '便利じゃん！でも今はFlutterじゃないんだろ？何があった？',
-      },
-      {
-        id: 'msg-003-4',
-        characterId: 'mill',
-        content: '実は本番環境で大きな問題が起きてね...\nログインがうまくいかなくなったっていうわけ。',
-      },
-      {
-        id: 'msg-003-5',
-        characterId: 'kettle',
-        content: 'ログインできなくなった？それヤバいじゃん！',
-      },
-      {
-        id: 'msg-003-6',
-        characterId: 'mill',
-        content: 'そうなんだ。開発中のテスト環境では動いてたのに、実際にユーザーが使う本番環境では動かなくなっちゃって...',
-      },
-      {
-        id: 'msg-003-7',
-        characterId: 'kettle',
-        content: '原因はわかったのか？',
-      },
-      {
-        id: 'msg-003-8',
-        characterId: 'mill',
-        content: '正直、完全には特定できなかったんだ。\n何度も修正を試みたけど、直してもまた別の問題が出てきてね...',
-      },
-      {
-        id: 'msg-003-9',
-        characterId: 'kettle',
-        content: 'もぐらたたき状態か。それは厄介だな。',
-      },
-      {
-        id: 'msg-003-10',
-        characterId: 'mill',
-        content: 'まさにそう！\nFlutterとFirebase（データを保存するサービス）の認証の組み合わせが複雑で、どこで問題が起きてるか追跡しづらかったんだ。',
-      },
-      {
-        id: 'msg-003-11',
-        characterId: 'kettle',
-        content: 'で、作り直す決断をしたわけだ？',
-      },
-      {
-        id: 'msg-003-12',
-        characterId: 'mill',
-        content: 'そうなんだ。修正し続けるより、最初から作り直した方が早いし確実だと判断したっていうわけ。\n苦渋の決断だったけどね。',
-      },
-      {
-        id: 'msg-003-13',
-        characterId: 'kettle',
-        content: '新しく選んだNext.jsってのは何だ？',
-      },
-      {
-        id: 'msg-003-14',
-        characterId: 'mill',
-        content: 'Next.jsは、ウェブアプリを作るためのフレームワークなんだ。\nReact（リアクト）っていう有名な技術をベースにしてるんだよ。',
-      },
-      {
-        id: 'msg-003-15',
-        characterId: 'kettle',
-        content: 'なんでFlutterじゃなくてNext.jsにしたんだ？',
-      },
-      {
-        id: 'msg-003-16',
-        characterId: 'mill',
-        content: 'いくつか理由があるんだ。\nまず、ウェブ技術の方が情報が多くて、問題が起きても解決しやすいっていうわけ。',
-      },
-      {
-        id: 'msg-003-17',
-        characterId: 'mill',
-        content: 'それと、Flutterはアプリストアからダウンロードが必要だけど、ウェブアプリならURLにアクセスするだけで使えるんだ。',
-      },
-      {
-        id: 'msg-003-18',
-        characterId: 'kettle',
-        content: 'アプリダウンロード不要ってのはいいな！手軽だ。',
-      },
-      {
-        id: 'msg-003-19',
-        characterId: 'mill',
-        content: 'それにね、Next.jsとFirebaseの組み合わせは認証の仕組みがシンプルで、トラブルが起きにくいんだよ。',
-      },
-      {
-        id: 'msg-003-20',
-        characterId: 'kettle',
-        content: '他にも選択肢はあったのか？',
-      },
-      {
-        id: 'msg-003-21',
-        characterId: 'mill',
-        content: 'もちろん！React Native（リアクトネイティブ）とか、Flutter Webとか色々検討したんだ。\nでも、ローストプラスの使い方を考えると、ウェブアプリが一番合ってたっていうわけ。',
-      },
-      {
-        id: 'msg-003-22',
-        characterId: 'kettle',
-        content: 'ユーザーにとって何が変わった？',
-      },
-      {
-        id: 'msg-003-23',
-        characterId: 'mill',
-        content: 'まず、ログインが安定したんだ！\n以前みたいに「あれ、ログインできない...」っていうストレスがなくなったっていうわけ。',
-      },
-      {
-        id: 'msg-003-24',
-        characterId: 'kettle',
-        content: 'それは重要だな！他には？',
-      },
-      {
-        id: 'msg-003-25',
-        characterId: 'mill',
-        content: 'ページの読み込みも速くなったんだ。\nNext.jsは表示速度の最適化が得意なフレームワークなんだよ。',
-      },
-      {
-        id: 'msg-003-26',
-        characterId: 'kettle',
-        content: 'でも作り直すのは大変だっただろ？',
-      },
-      {
-        id: 'msg-003-27',
-        characterId: 'mill',
-        content: '正直、かなり大変だったんだ。\n一からデザインを考え直して、機能も全部実装し直したからね。',
-      },
-      {
-        id: 'msg-003-28',
-        characterId: 'mill',
-        content: 'でも、バイブコーディングのおかげで思ったより早く形になったんだ。\nAIと一緒に開発できる時代でよかったよ。',
-      },
-      {
-        id: 'msg-003-29',
-        characterId: 'kettle',
-        content: 'この失敗から学んだことは何だ？',
-      },
-      {
-        id: 'msg-003-30',
-        characterId: 'mill',
-        content: 'たくさんあるんだ。\nまず、本番環境でのテストをもっと早い段階でやるべきだったっていうこと。',
-      },
-      {
-        id: 'msg-003-31',
-        characterId: 'mill',
-        content: 'それと、技術選定は「作りやすさ」だけじゃなく「問題が起きたときの対処しやすさ」も大事だって学んだんだ。',
-      },
-      {
-        id: 'msg-003-32',
-        characterId: 'kettle',
-        content: '失敗を次に活かしてるじゃん。それが大事だ。',
-      },
-      {
-        id: 'msg-003-33',
-        characterId: 'mill',
-        content: 'ユーザーの皆さんには迷惑をかけてしまって申し訳なかったけど、結果的により良いアプリになったと思うんだ。',
-      },
-      {
-        id: 'msg-003-34',
-        characterId: 'kettle',
-        content: '素直に失敗認めて改善する。それができてるなら問題ないだろ。',
-      },
-      {
-        id: 'msg-003-35',
-        characterId: 'mill',
-        content: 'ありがとう。これからも安定して使えるアプリを目指して頑張るよ！\n使ってくれるみんなに感謝なんだ。',
-      },
-    ],
-    detailContent: `ローストプラスは、一度Flutterで開発した後、Next.jsで作り直しています。この記事では、その背景と経緯を詳しく説明します。
-
-## Flutterで何が起きたのか
-
-### 本番環境でのログイン問題
-
-Flutterで開発したアプリをリリースした後、**本番環境でログインができなくなる問題**が発生しました。
-
-- 開発中のテスト環境では正常に動作していた
-- 実際のユーザーが使う本番環境でのみ問題が発生
-- ログイン処理が途中で止まったり、認証状態が保持されなかったりした
-
-### 原因の特定困難性
-
-問題の根本原因を完全に特定することはできませんでした。考えられる要因として：
-
-- FlutterとFirebase Authenticationの複雑な連携
-- 本番環境と開発環境の微妙な設定差異
-- 認証状態の管理に関する複数のレイヤー
-
-**何度も修正を試みましたが、1つ直すと別の問題が発生するという「もぐらたたき」状態**に陥りました。
-
-## なぜNext.jsを選んだのか
-
-### フレームワーク選定の理由
-
-| 比較項目 | Flutter | Next.js |
-|----------|---------|---------|
-| **情報量** | 比較的少ない | 非常に多い |
-| **認証連携** | 複雑 | シンプル |
-| **インストール** | アプリストア経由 | URL直接アクセス |
-| **更新反映** | ストア審査が必要 | 即時反映可能 |
-| **デバッグ** | やや困難 | ブラウザで容易 |
-
-### 他の選択肢との比較
-
-以下のフレームワークも検討しました：
-
-- **React Native**: アプリ形式を維持できるが、同様の認証問題リスク
-- **Flutter Web**: ウェブ対応だが、モバイル向け設計が影響
-- **Vue.js / Nuxt**: 良い選択肢だが、エコシステムの規模でNext.jsを選択
-
-**ローストプラスの主な使用シーン（店舗での利用）を考慮し、インストール不要なウェブアプリが最適と判断**しました。
-
-### 作り直しの決断
-
-修正を続けるか、作り直すかの判断基準：
-
-- ✅ 修正の見通しが立たない
-- ✅ 根本原因が特定できない
-- ✅ 同じフレームワークを使い続けるリスク
-- ✅ 新しい技術でより良い設計が可能
-
-これらの理由から、**一から作り直す判断**に至りました。
-
-## ユーザーへの影響
-
-### 改善されたポイント
-
-- **ログインの安定性**: 認証処理がシンプルになり、ログイン問題が解消
-- **読み込み速度**: Next.jsの最適化により、ページ表示が高速化
-- **アクセス性**: ブラウザから直接アクセス可能（アプリダウンロード不要）
-- **更新の即時反映**: 機能追加や修正がすぐにユーザーに届く
-
-### ユーザーが気付く変化
-
-- URLにアクセスするだけで使える
-- ログイン時のエラーや待ち時間が減少
-- 画面遷移がスムーズに
-
-## 失敗から学んだこと
-
-### 1. 早期の本番テスト
-
-開発の後半になってから本番環境でテストするのではなく、**早い段階で実際の環境に近い形で検証する**ことの重要性を学びました。
-
-開発中のテスト環境では問題なく動いていたため、本番環境特有の問題に気付くのが遅れました。もっと早い段階で本番環境に近い条件でテストしていれば、問題を早期に発見し、対処する時間を確保できたはずです。
-
-### 2. 技術選定の視点
-
-フレームワークやツールを選ぶとき、「作りやすさ」だけでなく**「問題が起きたときの対処しやすさ」**も重要な判断基準だと実感しました。
-
-- **情報の多さ**: 困ったときにネットで解決策が見つかるか
-- **コミュニティの活発さ**: 同じ問題に遭遇した人がいるか
-- **デバッグのしやすさ**: エラーの原因を追跡できる仕組みがあるか
-
-どんな技術にも長所・短所がありますが、トラブル発生時にどれだけ対処しやすいかは、プロジェクトの成功に大きく影響します。
-
-### 3. 損切りの判断
-
-修正を続けることと、作り直すことの**判断タイミング**を見極める重要性を学びました。
-
-「もう少し頑張れば直るかも」という思いで修正を続けがちですが、根本原因が不明なまま対症療法を繰り返すと、かえって時間を浪費することがあります。
-
-作り直すという決断は勇気がいりますが、結果的にはより早く、より良いものを作ることにつながりました。
-
-## ユーザーへの感謝
-
-Flutterバージョンで不便をおかけしたユーザーの皆様には、心よりお詫び申し上げます。
-
-この経験を糧に、より安定したサービスを提供できるよう努めていきます。引き続きローストプラスをよろしくお願いいたします。`,
-    tags: ['開発背景'],
-    publishedAt: '2026-01-17',
-    order: 3,
-  },
-  {
-    id: 'episode-004',
-    title: 'AIの今と未来',
-    subtitle: 'ChatGPTからAIエージェントへ、急速に進化するAIの世界',
-    imageUrl: '/dev-stories/ai_future.png',
-    dialogues: [
-      // 導入（5メッセージ）
-      {
-        id: 'msg-004-1',
-        characterId: 'press',
-        content: 'AIって最近よく聞くけど、要するにどのくらいすごくなってるの？',
-      },
-      {
-        id: 'msg-004-2',
-        characterId: 'siphon',
-        content: '実は2022年から2025年のたった3年で、想像を超える進化を遂げているんだよね。',
-      },
-      {
-        id: 'msg-004-3',
-        characterId: 'press',
-        content: '3年でそこまで変わるのか。テクノロジーって速いな。',
-      },
-      {
-        id: 'msg-004-4',
-        characterId: 'siphon',
-        content: '特に「AIエージェント」の登場で、AIの使い方が根本から変わってきているんだよね。',
-      },
-      {
-        id: 'msg-004-5',
-        characterId: 'press',
-        content: 'AIエージェント？それ何？',
-      },
-      // 主要AIの比較（10メッセージ）
-      {
-        id: 'msg-004-6',
-        characterId: 'siphon',
-        content: 'まず今のAIの全体像から説明するね。代表的なのはChatGPT、Gemini、Claudeの3つなんだよね。',
-      },
-      {
-        id: 'msg-004-7',
-        characterId: 'press',
-        content: 'ChatGPTは知ってる！他の2つは何が違うの？',
-      },
-      {
-        id: 'msg-004-8',
-        characterId: 'siphon',
-        content: 'ChatGPTはOpenAIという会社が作っていて、2022年に登場して一気に広まったんだよね。',
-      },
-      {
-        id: 'msg-004-9',
-        characterId: 'press',
-        content: 'あ、それで「AI」が一気に流行語になったわけか。納得。',
-      },
-      {
-        id: 'msg-004-10',
-        characterId: 'siphon',
-        content: 'そうなんだよね。GeminiはGoogleが作っているAIで、検索と連携できるのが強みなんだ。',
-      },
-      {
-        id: 'msg-004-11',
-        characterId: 'press',
-        content: 'Google製か。検索と連携できるのは強力だな。',
-      },
-      {
-        id: 'msg-004-12',
-        characterId: 'siphon',
-        content: 'Claudeはこのローストプラスを作るのにも使っているAIなんだよね。長い文章の理解が得意なんだ。',
-      },
-      {
-        id: 'msg-004-13',
-        characterId: 'press',
-        content: 'なるほど、このアプリもそのClaudeで作られてるわけか。',
-      },
-      {
-        id: 'msg-004-14',
-        characterId: 'siphon',
-        content: 'うん。どのAIも得意なことが少しずつ違うから、使い分けると便利なんだよね。',
-      },
-      {
-        id: 'msg-004-15',
-        characterId: 'press',
-        content: '使い分けか。でも普通の人にとっては、どれも似たようなものに見えるだろうな。',
-      },
-      // AIエージェントの登場（23メッセージ）
-      {
-        id: 'msg-004-16',
-        characterId: 'siphon',
-        content: '表面的にはそう見えるかもね。ただ最近は「AIエージェント」という新しい使い方が出てきて、これがすごいんだよね。',
-      },
-      {
-        id: 'msg-004-17',
-        characterId: 'press',
-        content: 'さっきも言ってたね。AIエージェントって結局何なの？',
-      },
-      {
-        id: 'msg-004-18',
-        characterId: 'siphon',
-        content: '2022年頃のAIは「チャット型」が主流だったんだよね。質問すると答えが返ってくる形式。',
-      },
-      {
-        id: 'msg-004-19',
-        characterId: 'press',
-        content: 'ああ、普通の会話みたいなやつな。今も基本はそれだよな。',
-      },
-      {
-        id: 'msg-004-20',
-        characterId: 'siphon',
-        content: 'でもプログラミングをするときは、AIが出したコードをわざわざコピーして貼り付ける必要があったんだよね。',
-      },
-      {
-        id: 'msg-004-21',
-        characterId: 'press',
-        content: 'それは確かに面倒だ。効率悪いな。',
-      },
-      {
-        id: 'msg-004-22',
-        characterId: 'siphon',
-        content: 'そうなんだよね。でも2025年になって「AIエージェント」が登場したんだ。',
-      },
-      {
-        id: 'msg-004-23',
-        characterId: 'press',
-        content: '2025年か。で、何がどう変わったんだ？',
-      },
-      {
-        id: 'msg-004-24',
-        characterId: 'siphon',
-        content: 'AIエージェントは、自分でファイルを読んで、コードを書いて、テストまで自動でやってくれるんだよね。',
-      },
-      {
-        id: 'msg-004-25',
-        characterId: 'press',
-        content: 'マジか、全部自動でやるのか。それは革命的だな。',
-      },
-      {
-        id: 'msg-004-26',
-        characterId: 'siphon',
-        content: '人間は「こういうものが作りたい」という指示を出す役割があるんだよね。実際の作業はAIがやってくれる。',
-      },
-      {
-        id: 'msg-004-27',
-        characterId: 'press',
-        content: 'どんなツールがあるの？',
-      },
-      {
-        id: 'msg-004-28',
-        characterId: 'siphon',
-        content: '有名なのはCursor、Claude Code、Codex、Antigravityあたりかな。',
-      },
-      {
-        id: 'msg-004-29',
-        characterId: 'press',
-        content: '4つもあるんでしょ？違いは何？',
-      },
-      {
-        id: 'msg-004-30',
-        characterId: 'siphon',
-        content: 'CursorはAIエディタのパイオニアで、プログラミング用のエディタにAIが組み込まれているんだよね。',
-      },
-      {
-        id: 'msg-004-31',
-        characterId: 'press',
-        content: 'エディタってのは、要するにプログラムを書くためのソフトってことか。',
-      },
-      {
-        id: 'msg-004-32',
-        characterId: 'siphon',
-        content: 'プログラムを書くためのメモ帳みたいなソフトなんだよね。そこにAIがいて、一緒に開発できる。',
-      },
-      {
-        id: 'msg-004-33',
-        characterId: 'press',
-        content: '一緒に開発できるのはいいな。他のツールは？',
-      },
-      {
-        id: 'msg-004-34',
-        characterId: 'siphon',
-        content: 'Claude Codeはターミナルという黒い画面で動くんだけど、自分でファイルを探して、自分で編集してくれるんだよね。',
-      },
-      {
-        id: 'msg-004-35',
-        characterId: 'press',
-        content: '黒い画面か。ハッカーっぽくてかっこいいな。',
-      },
-      {
-        id: 'msg-004-36',
-        characterId: 'siphon',
-        content: 'そう見えるかもね。CodexはOpenAIが作っていて、複雑な問題を解くのが得意。Antigravityは新しいツールで、フロントエンドの開発に強いんだよね。',
-      },
-      {
-        id: 'msg-004-37',
-        characterId: 'press',
-        content: 'ローストプラスはどれで作ったの？',
-      },
-      {
-        id: 'msg-004-38',
-        characterId: 'siphon',
-        content: '実はいろいろ使い分けているんだよね。メインはCursor、最近はClaude Codeにも挑戦中なんだ。',
-      },
-      // 今後の展望（18メッセージ）
-      {
-        id: 'msg-004-39',
-        characterId: 'press',
-        content: 'ここまで聞くと、これからもっとすごくなりそうだな。',
-      },
-      {
-        id: 'msg-004-40',
-        characterId: 'siphon',
-        content: '間違いなくね。実は今、AIの世界では「AGI」という言葉がすごく注目されているんだよね。',
-      },
-      {
-        id: 'msg-004-41',
-        characterId: 'press',
-        content: 'AGI？それ何？',
-      },
-      {
-        id: 'msg-004-42',
-        characterId: 'siphon',
-        content: 'AGIは「汎用人工知能」のこと。今のAIは特定のタスクが得意だけど、AGIは人間みたいに何でもできるAIなんだよね。',
-      },
-      {
-        id: 'msg-004-43',
-        characterId: 'press',
-        content: '何でもできるAIか。それが実現したらマジですごいな。',
-      },
-      {
-        id: 'msg-004-44',
-        characterId: 'siphon',
-        content: 'しかも、AIの進化は「指数関数的」に加速しているんだよね。1年ごとに2倍、3倍と速くなっていく感じ。',
-      },
-      {
-        id: 'msg-004-45',
-        characterId: 'press',
-        content: '指数関数的か。要するに加速し続けるってことだな。',
-      },
-      {
-        id: 'msg-004-46',
-        characterId: 'siphon',
-        content: '例えば、2022年から2025年の3年でこれだけ進化したでしょ？次の3年はもっと大きな変化が起きる可能性があるんだよね。',
-      },
-      {
-        id: 'msg-004-47',
-        characterId: 'press',
-        content: '次の3年でもっと変わるのか。想像つかないな。',
-      },
-      {
-        id: 'msg-004-48',
-        characterId: 'siphon',
-        content: 'そうなんだよね。これを「知能爆発」って呼ぶ人もいる。AIがAIを改良して、どんどん賢くなっていく。',
-      },
-      {
-        id: 'msg-004-49',
-        characterId: 'press',
-        content: 'AIがAIを改良するって、まさにSF映画の世界だな。',
-      },
-      {
-        id: 'msg-004-50',
-        characterId: 'siphon',
-        content: 'そして、AIが人間の知能を超える瞬間を「シンギュラリティ（技術的特異点）」って呼ぶんだよね。',
-      },
-      {
-        id: 'msg-004-51',
-        characterId: 'press',
-        content: 'シンギュラリティか、聞いたことあるな。2045年とか言われてたやつ。',
-      },
-      {
-        id: 'msg-004-52',
-        characterId: 'siphon',
-        content: '2045年って言われていたけど、今の進化スピードを見ると、もっと早いかもしれないんだよね。',
-      },
-      {
-        id: 'msg-004-53',
-        characterId: 'press',
-        content: 'ワクワクするけど、ちょっと怖くない？',
-      },
-      {
-        id: 'msg-004-54',
-        characterId: 'siphon',
-        content: 'わかるよ。でも大事なのは、AIに全部任せきりにしないこと。一緒に作っていく姿勢が大切だと思うんだよね。',
-      },
-      {
-        id: 'msg-004-55',
-        characterId: 'press',
-        content: '要するにAIと人間のコラボか。それなら納得だ。',
-      },
-      {
-        id: 'msg-004-56',
-        characterId: 'siphon',
-        content: 'そう！プログラミングを知らなくても、アイデアを伝えるだけでアプリが作れる時代が近づいているんだよね。',
-      },
-      // まとめ（2メッセージ）
-      {
-        id: 'msg-004-57',
-        characterId: 'siphon',
-        content: 'AIの進化はまだまだ続く。僕たちは今、歴史的な転換点にいるんだよね。これからも楽しみだね。',
-      },
-      {
-        id: 'msg-004-58',
-        characterId: 'press',
-        content: 'すごい時代に生きてるってことだな。これは楽しみだ！',
-      },
-    ],
-    detailContent: `AIは2022年から2025年の3年間で劇的に進化しました。この記事では、主要なAIの特徴からAIエージェントの登場、そして今後の展望までを解説します。
-
-## 主要AI（チャット型AI）の比較
-
-現在、代表的なAIとして**ChatGPT**、**Gemini**、**Claude**の3つがあります。
-
-| AI | 開発元 | 登場時期 | 特徴 |
-|-----|--------|----------|------|
-| **ChatGPT** | OpenAI | 2022年11月 | 最初に広まったAI、汎用性が高い |
-| **Gemini** | Google | 2023年12月 | Google検索と連携、マルチモーダル対応 |
-| **Claude** | Anthropic | 2023年3月 | 長文理解が得意、安全性を重視 |
-
-### それぞれの強み
-
-- **ChatGPT**: 幅広いタスクに対応でき、プラグインやGPTsで機能拡張が可能
-- **Gemini**: Googleの検索データと連携し、最新情報に強い。画像・動画・音声も扱える
-- **Claude**: 長い文脈を理解し、丁寧で正確な回答が得意。このローストプラスの開発にも使用
-
-## チャット型AIからAIエージェントへの進化
-
-### 2022年：チャット型AIの時代
-
-\`\`\`
-[ユーザー] → 質問する → [AI] → 回答を返す → [ユーザー] → コピペして使う
-\`\`\`
-
-- 質問すると回答が返ってくる
-- コードを書いてもらっても、**手動でコピー＆ペースト**が必要
-- ファイルの読み書きはAIにはできない
-
-### 2025年：AIエージェントの時代
-
-**AIエージェントの流れ**
-
-1. ユーザーが指示を出す
-2. AIエージェントが自律的に作業
-   - ファイルを読む
-   - コードを書く
-   - テストを実行
-   - エラーを修正
-3. 完成！
-
-- **自律的に**ファイルを読み書き
-- テストの実行やエラー修正も自動
-- 人間は「何を作りたいか」を伝える役割
-
-## AIエージェントツール比較
-
-| ツール | 特徴 | 得意なこと |
-|--------|------|-----------|
-| **Cursor** | AIエディタのパイオニア | エディタ内でAIと対話しながら開発 |
-| **Claude Code** | ターミナルベースで動作 | 自律的なファイル操作と開発 |
-| **Codex** | OpenAI製 | 高い推論力で複雑な問題を解決 |
-| **Antigravity** | 新興のAI IDE | フロントエンド開発に強い |
-
-### ローストプラスでの使い分け
-
-このアプリは複数のAIエージェントを使い分けて開発されています：
-
-- **Cursor**: メインの開発ツールとして日常的に使用
-- **Claude Code**: 最近挑戦中の新しいツール
-- **Codex**: 複雑なバックエンド機能の修正
-- **Antigravity**: フロントエンドの改善
-
-## 2022年→2025年の変化まとめ
-
-| 項目 | 2022年 | 2025年 |
-|------|--------|--------|
-| **AIの役割** | 質問に答える | 自律的に作業する |
-| **コード生成** | コピペが必要 | 直接ファイルに書き込む |
-| **開発スタイル** | 人間がすべて管理 | AIと協業 |
-| **必要なスキル** | プログラミング必須 | 指示出し力が重要 |
-
-## 今後の展望
-
-### AGI（汎用人工知能）とは
-
-**AGI（Artificial General Intelligence）** は「汎用人工知能」と訳され、人間のように**あらゆるタスクをこなせるAI**のことです。
-
-| 現在のAI | AGI |
-|----------|-----|
-| 特定のタスクに特化 | あらゆるタスクに対応 |
-| 学習したことだけできる | 新しいことも自分で学ぶ |
-| ツールとして使う | パートナーとして協業 |
-
-現在のAIは「特化型AI」と呼ばれ、文章生成やコード作成など特定の分野で力を発揮します。一方、AGIは人間のように創造性や判断力を持ち、未知の問題にも対応できます。
-
-### 知能爆発と指数関数的進化
-
-AIの進化は**指数関数的**に加速しています。
-
-\`\`\`
-2022年 → 2025年: 大きな進化
-2025年 → 2028年: さらに大きな進化
-2028年 → 2031年: 想像を超える進化？
-\`\`\`
-
-**知能爆発（Intelligence Explosion）** とは、AIが自分自身を改良できるようになることで、進化のスピードが爆発的に加速する現象です。
-
-- AIがAIを設計・改良する
-- 改良されたAIがさらに優れたAIを作る
-- このサイクルが加速度的に繰り返される
-
-### シンギュラリティ（技術的特異点）
-
-**シンギュラリティ（Singularity）** は、AIが人間の知能を超える瞬間を指します。
-
-- 未来学者レイ・カーツワイルが**2045年**に到来すると予測
-- しかし最近のAI進化のスピードを見ると、**もっと早い可能性**も
-- シンギュラリティ後の世界は予測不可能とも言われる
-
-### AIはさらに賢くなる
-
-- **推論能力の向上**: より複雑な問題を解けるように
-- **マルチモーダル化**: テキストだけでなく、画像・音声・動画も扱える
-- **長期記憶**: 過去のやり取りを覚えて、より良い提案ができる
-- **自律性の向上**: より少ない指示で、より大きなタスクをこなせる
-
-### 非エンジニアでもアプリが作れる時代
-
-プログラミングの知識がなくても、**アイデアを言葉で伝えるだけ**でアプリが作れる時代が近づいています。
-
-- 「こんなアプリが欲しい」と伝える
-- AIが設計・実装・テストまで行う
-- 人間はフィードバックと最終確認
-
-### 大切なこと
-
-AIが進化しても、大切なのは**AIに全部任せきりにしないこと**です。
-
-- AIの提案を鵜呑みにせず、確認する
-- 人間とAIの**コラボレーション**を意識する
-- 新しい技術を楽しみながら学ぶ姿勢
-- AIと共に成長していく心構え
-
-## まとめ
-
-AIは「答えを返すだけ」から「自律的に作業する」へと進化しています。ChatGPT、Gemini、Claudeといったチャット型AIから、Cursor、Claude Code、Codex、AntigravityといったAIエージェントへ。
-
-そして今、AGI、知能爆発、シンギュラリティといった言葉が現実味を帯びてきています。私たちは歴史的な転換点に立っているのかもしれません。
-
-この進化はまだまだ続きます。ワクワクしながら、AIと一緒に未来を作っていきましょう！`,
-    tags: ['AI', 'ChatGPT', 'Claude', 'Gemini', 'AIエージェント', 'AGI', 'シンギュラリティ', '技術解説'],
-    publishedAt: '2026-01-17',
-    order: 4,
-  },
-  {
-    id: 'episode-005',
-    title: 'いろんな資格取得に挑戦中！',
-    subtitle: 'IT・AI資格を楽しく1ヶ月集中で取得',
-    imageUrl: '/dev-stories/qualifications_challenge.png',
-    dialogues: [
-      // 導入
-      {
-        id: 'msg-005-1',
-        characterId: 'asairi',
-        content: 'ねえねえ、最近いろんな資格を取ってるって聞いたんだけど！',
-      },
-      {
-        id: 'msg-005-2',
-        characterId: 'fukairi',
-        content: 'うん、IT系とAI系の資格に挑戦してるんだ。\n実は生成AIパスポートとITパスポートに合格したんだよ。',
-      },
-      {
-        id: 'msg-005-3',
-        characterId: 'asairi',
-        content: 'すごい！おめでとう！他にも受けてるの？',
-      },
-      {
-        id: 'msg-005-4',
-        characterId: 'fukairi',
-        content: 'G検定も受けて、今は合格発表待ちなんだ。\nあと2月22日に基本情報技術者試験を受ける予定だよ。',
-      },
-      // 生成AIパスポート
-      {
-        id: 'msg-005-5',
-        characterId: 'asairi',
-        content: '生成AIパスポートって何？どんな資格なの？',
-      },
-      {
-        id: 'msg-005-6',
-        characterId: 'fukairi',
-        content: '生成AIを活用するための知識を証明する資格なんだ。\nChatGPTとかの仕組みや、ビジネス活用、著作権の問題なんかを学べるよ。',
-      },
-      {
-        id: 'msg-005-7',
-        characterId: 'asairi',
-        content: '今まさに必要な知識だね！どうやって勉強したの？',
-      },
-      {
-        id: 'msg-005-8',
-        characterId: 'fukairi',
-        content: '「生成AIパスポートテキスト&問題集」と「生成AIパスポート、トレ速」っていうアプリを使ったよ。\n普段からAIを使ってるから、実感を持って勉強できたんだ。',
-      },
-      {
-        id: 'msg-005-9',
-        characterId: 'asairi',
-        content: 'ローストプラスの開発でAIを使ってるもんね！',
-      },
-      {
-        id: 'msg-005-10',
-        characterId: 'fukairi',
-        content: 'そうなんだ。実際に使っている経験があると、テキストの内容がスッと入ってくるんだよね。',
-      },
-      // ITパスポート
-      {
-        id: 'msg-005-11',
-        characterId: 'asairi',
-        content: 'ITパスポートはどんな資格？',
-      },
-      {
-        id: 'msg-005-12',
-        characterId: 'fukairi',
-        content: 'ITの基礎知識を幅広く学べる国家資格だよ。\n経営戦略、マネジメント、テクノロジーの3分野から出題されるんだ。',
-      },
-      {
-        id: 'msg-005-13',
-        characterId: 'asairi',
-        content: '国家資格なんだ！難しそう...',
-      },
-      {
-        id: 'msg-005-14',
-        characterId: 'fukairi',
-        content: '実は思ったより取り組みやすかったよ。\n「過去問道場」っていう無料のサイトと、アプリ、あとAnkiっていう分散学習ツールを使ったんだ。',
-      },
-      {
-        id: 'msg-005-15',
-        characterId: 'asairi',
-        content: 'Anki？聞いたことないな。どんなツールなの？',
-      },
-      {
-        id: 'msg-005-16',
-        characterId: 'fukairi',
-        content: 'Ankiは忘れる頃に問題を出してくれるから、効率よく記憶に定着するんだ。\nスマホでどこでも勉強できるから、通勤中とか隙間時間に活用できたんだよ。',
-      },
-      // G検定
-      {
-        id: 'msg-005-17',
-        characterId: 'asairi',
-        content: 'G検定は？AIの資格だよね？',
-      },
-      {
-        id: 'msg-005-18',
-        characterId: 'fukairi',
-        content: 'そう、日本ディープラーニング協会が主催するAI・機械学習の検定だよ。\nディープラーニングの仕組みやビジネス活用について問われるんだ。',
-      },
-      {
-        id: 'msg-005-19',
-        characterId: 'asairi',
-        content: '難しそう...どんな勉強をしたの？',
-      },
-      {
-        id: 'msg-005-20',
-        characterId: 'fukairi',
-        content: '「ディープラーニングG検定（ジェネラリスト）最強の合格テキスト」と、AI問題集アプリを使ったよ。\n数学的な部分は難しかったけど、概念を理解することに集中したんだ。',
-      },
-      {
-        id: 'msg-005-21',
-        characterId: 'asairi',
-        content: '試験はどうだった？',
-      },
-      {
-        id: 'msg-005-22',
-        characterId: 'fukairi',
-        content: 'オンライン試験で、100分で145問を解くんだ。\n時間との戦いだったけど、なんとか最後まで解けたよ。',
-      },
-      {
-        id: 'msg-005-23',
-        characterId: 'asairi',
-        content: '145問！？それはすごいね！結果が楽しみだね。',
-      },
-      // 基本情報技術者試験
-      {
-        id: 'msg-005-24',
-        characterId: 'asairi',
-        content: '次は基本情報技術者試験を受けるんだよね？',
-      },
-      {
-        id: 'msg-005-25',
-        characterId: 'fukairi',
-        content: 'うん、2月22日に受験予定なんだ。\nITエンジニアの登竜門と言われる国家資格だよ。',
-      },
-      {
-        id: 'msg-005-26',
-        characterId: 'asairi',
-        content: 'ITパスポートより難しいの？',
-      },
-      {
-        id: 'msg-005-27',
-        characterId: 'fukairi',
-        content: 'うん、かなりレベルアップするよ。\nアルゴリズムやプログラミング、ネットワークなど、より専門的な内容が出題されるんだ。',
-      },
-      {
-        id: 'msg-005-28',
-        characterId: 'asairi',
-        content: '今どんな勉強をしてるの？',
-      },
-      {
-        id: 'msg-005-29',
-        characterId: 'fukairi',
-        content: 'かやのき先生の参考書で基礎を固めて、過去問道場で問題演習をしてるよ。\n科目A・科目Bの両方をバランスよく対策中なんだ。',
-      },
-      {
-        id: 'msg-005-30',
-        characterId: 'asairi',
-        content: '応援してるね！頑張って！',
-      },
-      // 勉強のコツ
-      {
-        id: 'msg-005-31',
-        characterId: 'asairi',
-        content: 'こんなにたくさんの資格を取るコツってあるの？',
-      },
-      {
-        id: 'msg-005-32',
-        characterId: 'fukairi',
-        content: '一番大事なのは「楽しむこと」かな。\n新しいことを学ぶのって、本来ワクワクするものだと思うんだ。',
-      },
-      {
-        id: 'msg-005-33',
-        characterId: 'asairi',
-        content: '勉強が楽しいってすごいね！',
-      },
-      {
-        id: 'msg-005-34',
-        characterId: 'fukairi',
-        content: 'あとは「短期集中」がポイントかな。\n長期間ダラダラやるより、約1ヶ月でギュッと集中した方が効率がいいんだ。',
-      },
-      {
-        id: 'msg-005-35',
-        characterId: 'asairi',
-        content: '確かに、モチベーションも維持しやすいよね！',
-      },
-      {
-        id: 'msg-005-36',
-        characterId: 'fukairi',
-        content: 'それと、スマホアプリや過去問サイトを活用すると、スキマ時間を有効に使えるよ。\n電車の中とか、ちょっとした待ち時間でも勉強できるんだ。',
-      },
-      // まとめ
-      {
-        id: 'msg-005-37',
-        characterId: 'asairi',
-        content: '資格を取ることで何か変わった？',
-      },
-      {
-        id: 'msg-005-38',
-        characterId: 'fukairi',
-        content: '知識が体系的に整理されたのが大きいかな。\nなんとなく使ってたAIやITの仕組みを、ちゃんと理解できるようになったんだ。',
-      },
-      {
-        id: 'msg-005-39',
-        characterId: 'asairi',
-        content: 'ローストプラスの開発にも活かせそう？',
-      },
-      {
-        id: 'msg-005-40',
-        characterId: 'fukairi',
-        content: 'もちろん！セキュリティやデータベースの知識とか、実際の開発に役立ってるよ。\n資格の勉強は「使える知識」を身につけるのにとても良い方法だと思うんだ。',
-      },
-      {
-        id: 'msg-005-41',
-        characterId: 'asairi',
-        content: 'これから資格を取ろうとしている人へメッセージはある？',
-      },
-      {
-        id: 'msg-005-42',
-        characterId: 'fukairi',
-        content: '完璧を目指さなくていいから、まず始めてみることが大事だよ。\n楽しみながら、自分のペースで挑戦してほしいな。一緒に頑張ろう！',
-      },
-    ],
-    detailContent: `IT・AI関連の資格取得に挑戦した体験談をお届けします。楽しく効率的に資格を取得するコツも紹介します！
-
-## 取得・挑戦中の資格一覧
-
-| 資格名 | 状況 | 試験形式 |
-|--------|------|----------|
-| **生成AIパスポート** | ✅ 合格 | オンライン |
-| **ITパスポート** | ✅ 合格 | CBT（テストセンター） |
-| **G検定** | ⏳ 合格発表待ち | オンライン |
-| **基本情報技術者試験** | 📚 勉強中（2月22日受験予定） | CBT（テストセンター） |
-
-## 生成AIパスポート
-
-### 資格の概要
-
-生成AIパスポートは、ChatGPTなどの生成AIを活用するための知識を証明する資格です。
-
-**出題範囲：**
-- 生成AIの仕組み（LLM、プロンプト、トランスフォーマーなど）
-- ビジネス活用と事例
-- 著作権・知的財産権の問題
-- AIの倫理とリスク
-- 最新のAI動向
-
-### 勉強方法
-
-- **「生成AIパスポートテキスト&問題集」**：基礎知識を体系的に学習
-- **「生成AIパスポート、トレ速」アプリ**：スマホで手軽に演習
-- **実践経験**：日常的にAIツールを使う
-
-### 合格のポイント
-
-普段からChatGPTやClaudeを使っていたので、実感を持って勉強できました。テキストの内容が「あ、これ知ってる！」と繋がる瞬間が多く、楽しく学べました。
-
-## ITパスポート
-
-### 資格の概要
-
-ITパスポートは、ITの基礎知識を幅広く問う国家資格です。
-
-**出題分野：**
-- **ストラテジ系**（経営戦略、マーケティング、法務など）
-- **マネジメント系**（プロジェクト管理、サービスマネジメントなど）
-- **テクノロジ系**（ネットワーク、セキュリティ、データベースなど）
-
-### 勉強方法
-
-- **過去問道場**（https://www.itpassportsiken.com/）：無料で過去問が解ける神サイト
-- **スマホアプリ**：通勤中や隙間時間に活用
-- **Anki**：分散学習ツール。忘れる頃に問題を出してくれるので効率よく記憶に定着する
-- **解説を読み込む**：間違えた問題は必ず理解する
-
-### 合格のポイント
-
-過去問道場とAnkiを徹底的に活用しました。電車の中、待ち時間、寝る前など、スマホがあればどこでも勉強できるのが良かったです。
-
-## G検定（ジェネラリスト検定）
-
-### 資格の概要
-
-G検定は、日本ディープラーニング協会（JDLA）が主催するAI・機械学習の検定です。
-
-**出題範囲：**
-- 人工知能の歴史と定義
-- 機械学習の手法（教師あり学習、教師なし学習など）
-- ディープラーニングの仕組み
-- CNNやRNNなどのニューラルネットワーク
-- AI活用の法律・倫理
-
-### 勉強方法
-
-- **「ディープラーニングG検定（ジェネラリスト）最強の合格テキスト」**：基礎をしっかり学習
-- **AI問題集アプリ**：スマホで手軽に演習
-- **過去問対策**：時間配分を意識して解く
-
-### 試験の感想
-
-100分で145問という問題数。最初は焦りましたが、サクサク解くことを意識して最後まで完走できました。合格発表が楽しみです！
-
-## 基本情報技術者試験（勉強中）
-
-### 資格の概要
-
-基本情報技術者試験は、ITエンジニアの登竜門と言われる国家資格です。
-
-**試験構成：**
-- **科目A**：60問（90分）- 知識問題
-- **科目B**：20問（100分）- アルゴリズム・情報セキュリティ
-
-### 勉強方法
-
-- **かやのき式 基本情報技術者**：わかりやすい解説で基礎を固める
-- **過去問道場**：科目A対策に最適
-- **アルゴリズム対策**：トレース練習を重点的に
-
-### 現在の進捗
-
-2月22日の受験に向けて追い込み中です。科目Aは過去問で8割以上取れるようになってきたので、科目Bのアルゴリズム対策に力を入れています。
-
-## 1ヶ月集中で資格を取るコツ
-
-### 1. 楽しむこと
-
-勉強を「やらなきゃいけないこと」ではなく「新しいことを知る楽しみ」と捉えると、モチベーションが続きます。
-
-### 2. 短期集中
-
-長期間ダラダラ勉強するより、約1ヶ月でギュッと集中した方が効率的です。
-
-- **ITパスポート**：約1ヶ月
-- **生成AIパスポート**：約1ヶ月
-- **G検定**：約1ヶ月
-
-### 3. スキマ時間の活用
-
-- 通勤時間
-- 待ち時間
-- 寝る前の10分
-
-スマホアプリや過去問サイトを使えば、どこでも勉強できます。
-
-### 4. 完璧を目指さない
-
-100%理解してから次に進むのではなく、まず全体像をつかんでから深掘りする方が効率的です。
-
-### 5. 実践と結びつける
-
-資格の勉強で得た知識を、実際のプロジェクトで使うと定着します。ローストプラスの開発でも、セキュリティやデータベースの知識が役立っています。
-
-## おすすめの勉強ツール
-
-| ツール | 対象資格 | 特徴 |
-|--------|----------|------|
-| **過去問道場** | ITパスポート、基本情報 | 無料で過去問が解ける |
-| **AI問題集アプリ** | G検定 | スマホで手軽に演習 |
-| **Anki** | 全般 | 分散学習で効率よく記憶に定着 |
-| **公式テキスト** | 各資格 | 体系的に学べる |
-| **YouTube** | 全般 | 解説動画で理解を深める |
-
-## これから資格を取る人へ
-
-資格取得は、知識を体系的に整理する良い機会です。
-
-- **始めることが大事**：完璧な準備を待たずに、まず始めてみる
-- **楽しむこと**：新しいことを学ぶワクワク感を大切に
-- **自分のペースで**：他人と比べず、自分に合った勉強法を見つける
-
-一緒に頑張りましょう！`,
-    tags: ['資格', 'IT', '勉強法', '体験談'],
-    publishedAt: '2026-01-24',
-    order: 5,
-  },
-  {
-    id: 'episode-006',
-    title: 'Claude Codeを使ってみた',
-    subtitle: '文字でAIと会話しながら開発できる新しいツール',
-    imageUrl: '/dev-stories/claude_code.png',
-    dialogues: [
-      // 導入（5メッセージ）
-      {
-        id: 'msg-006-1',
-        characterId: 'press',
-        content: '最近「Claude Code」っていうツールを使い始めたって聞いたけど、どういうものなの？',
-      },
-      {
-        id: 'msg-006-2',
-        characterId: 'siphon',
-        content: '簡単に言うと、文字で会話しながらアプリを作れるAIアシスタントなんだ。',
-      },
-      {
-        id: 'msg-006-3',
-        characterId: 'press',
-        content: 'Cursorとは違うの？',
-      },
-      {
-        id: 'msg-006-4',
-        characterId: 'siphon',
-        content: 'Cursorはエディタ画面の中でAIと対話するスタイル。Claude Codeはターミナルっていう黒い画面でテキストだけでやり取りするんだ。シンプルだけど、その分パワフルなんだよね。',
-      },
-      {
-        id: 'msg-006-5',
-        characterId: 'press',
-        content: 'なんだかレトロな感じがするけど、どうしてそっちを使うの？',
-      },
-      // AIアシスタントの種類（8メッセージ）
-      {
-        id: 'msg-006-6',
-        characterId: 'siphon',
-        content: '実はアプリ開発を手伝ってくれるAIにも色々なタイプがあるんだ。',
-      },
-      {
-        id: 'msg-006-7',
-        characterId: 'press',
-        content: 'どんな種類があるの？',
-      },
-      {
-        id: 'msg-006-8',
-        characterId: 'siphon',
-        content: '一つ目は「エディタ統合型」。CursorやAntigravityみたいに、コードエディタの中にAIが組み込まれているタイプだね。',
-      },
-      {
-        id: 'msg-006-9',
-        characterId: 'press',
-        content: 'それは前から使ってたよね。他には？',
-      },
-      {
-        id: 'msg-006-10',
-        characterId: 'siphon',
-        content: '二つ目が「ターミナル型」。Claude Codeがこれにあたるね。テキストだけでやり取りするから、動作が軽くて柔軟性が高いんだ。',
-      },
-      {
-        id: 'msg-006-11',
-        characterId: 'press',
-        content: 'コードを書かなくてもいいタイプもあるって聞いたけど？',
-      },
-      {
-        id: 'msg-006-12',
-        characterId: 'siphon',
-        content: 'LovableやBoltみたいに「こんなアプリが欲しい」って伝えるだけで作ってくれるツールもあるよ。ただ、細かい調整が難しいんだよね。',
-      },
-      {
-        id: 'msg-006-13',
-        characterId: 'press',
-        content: 'それぞれ得意なことが違うんだね。Claude Codeは何が得意なの？',
-      },
-      // Claude Codeの便利機能（12メッセージ）
-      {
-        id: 'msg-006-14',
-        characterId: 'siphon',
-        content: 'Claude Codeのすごいところは「自分で考えて動いてくれる」ことなんだ。例えば「この機能を直して」って言うと…',
-      },
-      {
-        id: 'msg-006-15',
-        characterId: 'press',
-        content: 'どうなるの？',
-      },
-      {
-        id: 'msg-006-16',
-        characterId: 'siphon',
-        content: 'まず自分でファイルを探して、中身を読んで、問題を見つけて、直すところまで全部やってくれるんだ。まるで優秀な助手がいるみたい。',
-      },
-      {
-        id: 'msg-006-17',
-        characterId: 'press',
-        content: 'それはすごい！他にも便利な機能はある？',
-      },
-      {
-        id: 'msg-006-18',
-        characterId: 'siphon',
-        content: '「スキル」っていう機能があって、よく使う作業を一言で頼めるんだ。例えば「/commit」って打つだけで、変更を記録してくれる。',
-      },
-      {
-        id: 'msg-006-19',
-        characterId: 'press',
-        content: '料理のレシピみたいに、定番の手順が登録されてるってこと？',
-      },
-      {
-        id: 'msg-006-20',
-        characterId: 'siphon',
-        content: 'そうそう！自分だけのオリジナルレシピも作れるんだよ。ローストプラスでは「UIデザインを確認する」スキルとか作ってるんだ。',
-      },
-      {
-        id: 'msg-006-21',
-        characterId: 'press',
-        content: '他のツールとも連携できるって聞いたけど？',
-      },
-      {
-        id: 'msg-006-22',
-        characterId: 'siphon',
-        content: 'うん、「外部ツール連携」っていう仕組みがあってね。ローストプラスの開発では3つ使ってるんだ。',
-      },
-      {
-        id: 'msg-006-23',
-        characterId: 'press',
-        content: 'どんなことができるの？',
-      },
-      {
-        id: 'msg-006-24',
-        characterId: 'siphon',
-        content: '一つ目はコードの中身を賢く探せる機能。二つ目は最新のマニュアルを自動で調べてくれる機能。三つ目はブラウザ画面を見て確認してくれる機能。',
-      },
-      {
-        id: 'msg-006-25',
-        characterId: 'press',
-        content: 'AIがブラウザ画面を見れるの？',
-      },
-      {
-        id: 'msg-006-26',
-        characterId: 'siphon',
-        content: 'そう！「このボタンをクリックして」って頼むと、実際に画面を見ながら操作してくれるんだ。表示がおかしくないかも確認できるよ。',
-      },
-      // 開発効率の向上（6メッセージ）
-      {
-        id: 'msg-006-27',
-        characterId: 'press',
-        content: 'ローストプラスの開発でどう役立ってるの？',
-      },
-      {
-        id: 'msg-006-28',
-        characterId: 'siphon',
-        content: '実はこの「開発秘話」ページ自体がClaude Codeで作られてるんだ。新しいエピソードを追加するのも、ほとんど自動でやってくれる。',
-      },
-      {
-        id: 'msg-006-29',
-        characterId: 'press',
-        content: 'へぇ！具体的にどう楽になったの？',
-      },
-      {
-        id: 'msg-006-30',
-        characterId: 'siphon',
-        content: '例えば「焙煎タイマーの表示を大きくして」って頼むと、関連するファイルを探して、コードを直して、画面で確認するところまでやってくれるんだ。',
-      },
-      {
-        id: 'msg-006-31',
-        characterId: 'press',
-        content: 'コーヒーを淹れてる間に作業が進んでそうだね。',
-      },
-      {
-        id: 'msg-006-32',
-        characterId: 'siphon',
-        content: 'まさにそんな感じ！以前は自分で全部やってたから、体感で3倍くらい速くなったかな。',
-      },
-      // Cursorとの比較（4メッセージ）
-      {
-        id: 'msg-006-33',
-        characterId: 'press',
-        content: 'じゃあCursorはもう使わないの？',
-      },
-      {
-        id: 'msg-006-34',
-        characterId: 'siphon',
-        content: 'いや、両方使い分けてるよ。Cursorは画面を見ながら細かく調整したい時に便利。Claude Codeは「この機能を作って」って大きく任せたい時に使うんだ。',
-      },
-      {
-        id: 'msg-006-35',
-        characterId: 'press',
-        content: 'コーヒーで例えると、どんな感じ？',
-      },
-      {
-        id: 'msg-006-36',
-        characterId: 'siphon',
-        content: 'Cursorはハンドドリップみたいに丁寧に一杯ずつ淹れる感じ。Claude Codeは優秀なバリスタに「美味しいコーヒーをお願い」って任せる感じかな。',
-      },
-      // まとめ（4メッセージ）
-      {
-        id: 'msg-006-37',
-        characterId: 'press',
-        content: 'AIと一緒に開発する時代が本格的に来たんだね。',
-      },
-      {
-        id: 'msg-006-38',
-        characterId: 'siphon',
-        content: 'そうだね。でも大事なのは「何を作りたいか」っていうアイデア。道具が便利になっても、そこは変わらないと思うんだ。',
-      },
-      {
-        id: 'msg-006-39',
-        characterId: 'press',
-        content: 'ローストプラスも、コーヒーへの愛情があってこそってことだね。',
-      },
-      {
-        id: 'msg-006-40',
-        characterId: 'siphon',
-        content: 'その通り！これからもAIの力を借りながら、もっと使いやすいアプリにしていきたいな。',
-      },
-    ],
-    detailContent: `Claude Codeは、文字で会話しながらアプリを作れるAIアシスタントです。この記事では、AIアシスタントの種類からClaude Codeの便利な機能、そしてローストプラスでどう活用しているかをわかりやすく紹介します。
-
-## 開発を手伝ってくれるAIアシスタントとは
-
-最近、アプリ開発を手伝ってくれるAIツールが増えています。「こんな機能を作って」とお願いすると、自分でファイルを探して、コードを書いて、動作確認まで手伝ってくれます。まるで優秀な助手がそばにいるような感覚です。
-
-## AIアシスタントの種類
-
-開発を手伝ってくれるAIツールには、いくつかのタイプがあります。
-
-| タイプ | 特徴 | 代表的なツール |
-|--------|------|--------------|
-| **エディタ統合型** | コードエディタにAIが組み込まれている | Cursor, Antigravity |
-| **ターミナル型** | テキストベースで柔軟に対話 | Claude Code, Codex |
-| **ノーコード型** | 自然言語だけでアプリを生成 | Lovable, Bolt |
-
-### エディタ統合型
-
-CursorやAntigravityは、コードエディタの中にAIが組み込まれています。
-
-- **良いところ**: 視覚的で分かりやすい、コードのハイライトや補完が充実
-- **苦手なところ**: エディタに依存する
-
-### ターミナル型
-
-Claude Codeは、ターミナル（黒い画面）でテキストベースでやり取りします。
-
-- **良いところ**: 動作が軽い、スクリプトとの連携がしやすい、拡張性が高い
-- **苦手なところ**: 最初は慣れが必要
-
-### ノーコード型
-
-LovableやBoltは、「こんなアプリが欲しい」と自然言語で伝えるだけで形にしてくれます。
-
-- **良いところ**: プログラミング経験がなくても始められる
-- **苦手なところ**: 細かいカスタマイズが難しい
-
-## Claude Codeの便利な機能
-
-### 「自分で考えて動いてくれる」
-
-Claude Codeの最大の特徴は、自律的に作業してくれること。「この機能を直して」とお願いすると：
-
-1. 関連するファイルを自分で探す
-2. 中身を読んで問題を見つける
-3. コードを直す
-4. 動作確認まで手伝ってくれる
-
-まるで優秀なアシスタントがそばにいるような体験ができます。
-
-### スキル機能 - 定番の作業を一言で
-
-よく使う作業を「レシピ」のように登録しておいて、一言で呼び出せる機能です。
-
-**最初から使えるもの：**
-- \`/commit\` - 変更を記録
-- \`/review-pr\` - 変更内容をチェック
-
-**ローストプラスで作ったオリジナル：**
-- \`/roastplus-ui\` - デザインのルールを確認
-- \`/nextjs-firestore\` - データベース連携の書き方を確認
-
-料理でいえば、「チャーハンを作って」と言うだけで、材料を出して、切って、炒めて、盛り付けまでやってくれるイメージです。
-
-### 外部ツール連携 - AIの能力を拡張
-
-Claude Codeは他のツールと連携して、もっと賢くなれます。ローストプラスでは3つの連携を使っています：
-
-| 連携ツール | 何ができる？ |
-|-----------|------------|
-| **Serena** | コードの中身を賢く探す。「この機能を使っている場所を全部見つけて」が可能 |
-| **Context7** | 最新のマニュアルを自動で調べる。「この書き方で合ってる？」を確認 |
-| **Chrome DevTools** | ブラウザ画面を見て確認。「このボタンをクリックして」が可能 |
-
-### おまかせモード - 信頼できる操作は自動で
-
-通常は「この変更をしていい？」と毎回確認されますが、信頼できる操作は自動でOKするように設定できます。
-
-**安心ポイント：**
-- 変更履歴は全部残るので、いつでも元に戻せる
-- 最初は慎重に、慣れてから範囲を広げる
-
-## ローストプラスでの活用例
-
-### この「開発秘話」ページを作った
-
-実はこの開発秘話ページ自体が、Claude Codeで作られています。
-
-**作業の流れ：**
-1. 「新しいエピソードを追加したい」とお願い
-2. Claude Codeがファイル構造を理解
-3. 新しいエピソードのコードを生成
-4. 画面で表示を確認
-5. 完成！
-
-### 機能の改善
-
-「焙煎タイマーの表示を大きくして」とお願いすると：
-- 関連するファイルを探す
-- デザインのコードを修正
-- 画面で確認して完了
-
-コーヒーを淹れている間に作業が進んでいることも。
-
-### デザインの確認
-
-ブラウザ連携を使って、実際の画面を見ながら開発できます。
-- 「この画面のスクリーンショットを撮って」
-- 「このボタンをクリックして動作確認して」
-- 「エラーが出てないか確認して」
-
-## CursorとClaude Codeの使い分け
-
-どちらも素晴らしいツールなので、場面で使い分けています。
-
-| 場面 | 使うツール | 理由 |
-|------|-----------|------|
-| 細かい調整 | Cursor | 画面を見ながら微調整しやすい |
-| 大きな機能追加 | Claude Code | 「作って」と任せられる |
-| デザイン確認 | Claude Code | ブラウザ連携が便利 |
-| 学習中 | Cursor | 視覚的で分かりやすい |
-
-**コーヒーで例えると：**
-- **Cursor**: ハンドドリップ。丁寧に一杯ずつ淹れる
-- **Claude Code**: 優秀なバリスタ。「美味しいコーヒーをお願い」と任せられる
-
-## まとめ
-
-Claude Codeは、文字で会話するシンプルなスタイルながら、自分で考えて動いてくれる優秀なAIアシスタントです。
-
-- **自律的**: ファイルを探して、読んで、直すまで自分でやってくれる
-- **レシピ機能**: よく使う作業を一言で頼める
-- **連携機能**: 他のツールと組み合わせてもっと賢く
-- **安心設計**: 変更履歴が残るので、いつでも元に戻せる
-
-AIと一緒に開発する時代が来ています。でも大切なのは、「何を作りたいか」というアイデアと情熱。ローストプラスも、コーヒーへの愛情があってこそ。これからもAIの力を借りながら、もっと使いやすいアプリを目指していきます。`,
-    tags: ['Claude Code', 'AI開発', 'ツール紹介', '開発効率'],
-    publishedAt: '2026-01-25',
-    order: 6,
-  },
+  EPISODE_001,
+  EPISODE_002,
+  EPISODE_003,
+  EPISODE_004,
+  EPISODE_005,
+  EPISODE_006,
 ];
 
 // エピソードをIDで取得
